fix for https://github.com/surveyjs/survey-creator/issues/1945 : ability to disable adaptivity; extract MatrixDropDownColumn and MatrixDropDownRendered classes
