+++ src/question_matrixdropdownbase.ts
@@ -1,36 +1,15 @@
-import {
-  JsonObject,
-  CustomPropertiesCollection,
-  JsonObjectProperty,
-  Serializer,
-  property
-} from "./jsonobject";
+import { JsonObject, CustomPropertiesCollection, Serializer } from "./jsonobject";
 import { QuestionMatrixBaseModel } from "./martixBase";
 import { Question } from "./question";
 import { HashTable, Helpers } from "./helpers";
 import { Base } from "./base";
-import {
-  IElement,
-  IQuestion,
-  ISurveyData,
-  ISurvey,
-  ISurveyImpl,
-  ITextProcessor,
-  IProgressInfo,
-  IPanel,
-  IWrapperObject
-} from "./base-interfaces";
+import { IElement, IQuestion, ISurveyData, ISurvey, ISurveyImpl, ITextProcessor, IProgressInfo, IPanel } from "./base-interfaces";
 import { SurveyElement } from "./survey-element";
-import {
-  TextPreProcessorValue,
-  QuestionTextProcessor
-} from "./textPreProcessor";
+import { TextPreProcessorValue, QuestionTextProcessor } from "./textPreProcessor";
 import { ItemValue } from "./itemvalue";
 import { surveyLocalization } from "./surveyStrings";
-import { QuestionSelectBase } from "./question_baseselect";
 import { QuestionFactory } from "./questionfactory";
 import { ILocalizableOwner, LocalizableString } from "./localizablestring";
-import { SurveyValidator } from "./validator";
 import { getCurrecyCodes } from "./question_expression";
 import { FunctionFactory } from "./functionsfactory";
 import { PanelModel } from "./panel";
@@ -38,9 +17,9 @@ import { settings } from "./settings";
 import { KeyDuplicationError } from "./error";
 import { SurveyModel } from "./survey";
 import { SurveyError } from "./survey-error";
-import { Action, IAction } from "./actions/action";
-import { AdaptiveActionContainer } from "./actions/adaptive-container";
 import { CssClassBuilder } from "./utils/cssClassBuilder";
+import { MatrixDropdownColumn } from "./question_matrixdropdowncolumn";
+import { QuestionMatrixDropdownRenderedCell, QuestionMatrixDropdownRenderedRow, QuestionMatrixDropdownRenderedTable } from "./question_matrixdropdownrendered";
 
 export interface IMatrixDropdownData {
   value: any;
@@ -88,620 +67,6 @@ export interface IMatrixDropdownData {
   getSurvey(): ISurvey;
 }
 
-export interface IMatrixColumnOwner extends ILocalizableOwner {
-  getRequiredText(): string;
-  onColumnPropertyChanged(
-    column: MatrixDropdownColumn,
-    name: string,
-    newValue: any
-  ): void;
-  onColumnItemValuePropertyChanged(
-    column: MatrixDropdownColumn,
-    propertyName: string,
-    obj: ItemValue,
-    name: string,
-    newValue: any,
-    oldValue: any
-  ): void;
-  onShowInMultipleColumnsChanged(column: MatrixDropdownColumn): void;
-  getCellType(): string;
-  getCustomCellType(column: MatrixDropdownColumn, row: MatrixDropdownRowModelBase, cellType: string): string;
-  onColumnCellTypeChanged(column: MatrixDropdownColumn): void;
-}
-
-function onUpdateSelectBaseCellQuestion(
-  cellQuestion: QuestionSelectBase,
-  column: MatrixDropdownColumn,
-  question: QuestionMatrixDropdownModelBase,
-  data: any
-) {
-  cellQuestion.storeOthersAsComment = !!question
-    ? question.storeOthersAsComment
-    : false;
-  if (
-    (!cellQuestion.choices || cellQuestion.choices.length == 0) &&
-    cellQuestion.choicesByUrl.isEmpty
-  ) {
-    cellQuestion.choices = question.choices;
-  }
-  if (!cellQuestion.choicesByUrl.isEmpty) {
-    cellQuestion.choicesByUrl.run(data.getTextProcessor());
-  }
-}
-export var matrixDropdownColumnTypes = {
-  dropdown: {
-    properties: [
-      "choices",
-      "choicesOrder",
-      "choicesByUrl",
-      "optionsCaption",
-      "otherText",
-      "choicesVisibleIf",
-    ],
-    onCellQuestionUpdate: (
-      cellQuestion: any,
-      column: any,
-      question: any,
-      data: any
-    ) => {
-      onUpdateSelectBaseCellQuestion(cellQuestion, column, question, data);
-      if (
-        !!cellQuestion.locOptionsCaption &&
-        cellQuestion.locOptionsCaption.isEmpty &&
-        !question.locOptionsCaption.isEmpty
-      ) {
-        cellQuestion.optionsCaption = question.optionsCaption;
-      }
-    },
-  },
-  checkbox: {
-    properties: [
-      "choices",
-      "choicesOrder",
-      "choicesByUrl",
-      "otherText",
-      "choicesVisibleIf",
-      "hasSelectAll",
-      "hasNone",
-    ],
-    onCellQuestionUpdate: (
-      cellQuestion: any,
-      column: any,
-      question: any,
-      data: any
-    ) => {
-      onUpdateSelectBaseCellQuestion(cellQuestion, column, question, data);
-      cellQuestion.colCount =
-        column.colCount > -1 ? column.colCount : question.columnColCount;
-    },
-  },
-  radiogroup: {
-    properties: [
-      "choices",
-      "choicesOrder",
-      "choicesByUrl",
-      "otherText",
-      "choicesVisibleIf",
-    ],
-    onCellQuestionUpdate: (
-      cellQuestion: any,
-      column: any,
-      question: any,
-      data: any
-    ) => {
-      onUpdateSelectBaseCellQuestion(cellQuestion, column, question, data);
-      cellQuestion.colCount =
-        column.colCount > -1 ? column.colCount : question.columnColCount;
-    },
-  },
-  text: {
-    properties: ["placeHolder", "inputType", "maxLength", "min", "max", "step"],
-    onCellQuestionUpdate: (
-      cellQuestion: any,
-      column: any,
-      question: any,
-      data: any
-    ) => {},
-  },
-  comment: {
-    properties: ["placeHolder", "rows", "maxLength"],
-    onCellQuestionUpdate: (
-      cellQuestion: any,
-      column: any,
-      question: any,
-      data: any
-    ) => {},
-  },
-  boolean: {
-    properties: ["renderAs", "defaultValue"],
-    onCellQuestionUpdate: (
-      cellQuestion: any,
-      column: any,
-      question: any,
-      data: any
-    ) => {
-      cellQuestion.showTitle = true;
-      cellQuestion.renderAs = column.renderAs;
-    },
-  },
-  expression: {
-    properties: ["expression", "displayStyle", "currency"],
-    onCellQuestionUpdate: (
-      cellQuestion: any,
-      column: any,
-      question: any,
-      data: any
-    ) => {},
-  },
-  rating: {
-    properties: ["rateValues"],
-  },
-};
-
-export class MatrixDropdownColumn extends Base
-  implements ILocalizableOwner, IWrapperObject {
-  public static getColumnTypes(): Array<string> {
-    var res = [];
-    for (var key in matrixDropdownColumnTypes) {
-      res.push(key);
-    }
-    return res;
-  }
-  private templateQuestionValue: Question;
-  private colOwnerValue: IMatrixColumnOwner = null;
-  private indexValue = -1;
-  private _isVisible = true;
-  private _hasVisibleCell = true;
-
-  constructor(name: string, title: string = null) {
-    super();
-    var self = this;
-    this.createLocalizableString("totalFormat", this);
-    this.registerFunctionOnPropertyValueChanged(
-      "showInMultipleColumns",
-      function() {
-        self.doShowInMultipleColumnsChanged();
-      }
-    );
-    this.updateTemplateQuestion();
-    this.name = name;
-    if (title) {
-      this.title = title;
-    } else {
-      this.templateQuestion.locTitle.strChanged();
-    }
-  }
-  public getOriginalObj(): Base {
-    return this.templateQuestion;
-  }
-  getClassNameProperty(): string {
-    return "cellType";
-  }
-  public getSurvey(live: boolean = false): ISurvey {
-    return !!this.colOwner ? (<any>this.colOwner).survey : null;
-  }
-  endLoadingFromJson() {
-    super.endLoadingFromJson();
-    this.templateQuestion.endLoadingFromJson();
-    this.templateQuestion.onGetSurvey = () => {
-      return this.getSurvey();
-    };
-  }
-  getDynamicPropertyName(): string {
-    return "cellType";
-  }
-  getDynamicType(): string {
-    return this.calcCellQuestionType(null);
-  }
-  public get colOwner(): IMatrixColumnOwner {
-    return this.colOwnerValue;
-  }
-  public set colOwner(value: IMatrixColumnOwner) {
-    this.colOwnerValue = value;
-    if (!!value) {
-      this.updateTemplateQuestion();
-    }
-  }
-  public locStrsChanged() {
-    super.locStrsChanged();
-    this.locTitle.strChanged();
-  }
-  public addUsedLocales(locales: Array<string>) {
-    super.addUsedLocales(locales);
-    this.templateQuestion.addUsedLocales(locales);
-  }
-  public get index() {
-    return this.indexValue;
-  }
-  public setIndex(val: number) {
-    this.indexValue = val;
-  }
-  public getType() {
-    return "matrixdropdowncolumn";
-  }
-  public get cellType(): string {
-    return this.getPropertyValue("cellType");
-  }
-  public set cellType(val: string) {
-    val = val.toLocaleLowerCase();
-    this.setPropertyValue("cellType", val);
-    this.updateTemplateQuestion();
-    if (!!this.colOwner) {
-      this.colOwner.onColumnCellTypeChanged(this);
-    }
-  }
-  public get templateQuestion() {
-    return this.templateQuestionValue;
-  }
-  public get value() {
-    return this.templateQuestion.name;
-  }
-  public get isVisible() {
-    return this._isVisible;
-  }
-  public setIsVisible(newVal: boolean) {
-    this._isVisible = newVal;
-  }
-  public get hasVisibleCell() {
-    return this._hasVisibleCell;
-  }
-  public set hasVisibleCell(newVal: boolean) {
-    this._hasVisibleCell = newVal;
-  }
-  public get name() {
-    return this.templateQuestion.name;
-  }
-  public set name(val: string) {
-    this.templateQuestion.name = val;
-  }
-  public get title(): string {
-    return this.templateQuestion.title;
-  }
-  public set title(val: string) {
-    this.templateQuestion.title = val;
-  }
-  public get locTitle() {
-    return this.templateQuestion.locTitle;
-  }
-  public get fullTitle(): string {
-    return this.locTitle.textOrHtml;
-  }
-  public get isRequired(): boolean {
-    return this.templateQuestion.isRequired;
-  }
-  public set isRequired(val: boolean) {
-    this.templateQuestion.isRequired = val;
-  }
-  public get requiredText(): string {
-    return this.templateQuestion.requiredText;
-  }
-  public get requiredErrorText(): string {
-    return this.templateQuestion.requiredErrorText;
-  }
-  public set requiredErrorText(val: string) {
-    this.templateQuestion.requiredErrorText = val;
-  }
-  get locRequiredErrorText(): LocalizableString {
-    return this.templateQuestion.locRequiredErrorText;
-  }
-  public get readOnly(): boolean {
-    return this.templateQuestion.readOnly;
-  }
-  public set readOnly(val: boolean) {
-    this.templateQuestion.readOnly = val;
-  }
-  public get hasOther(): boolean {
-    return this.templateQuestion.hasOther;
-  }
-  public set hasOther(val: boolean) {
-    this.templateQuestion.hasOther = val;
-  }
-  public get visibleIf(): string {
-    return this.templateQuestion.visibleIf;
-  }
-  public set visibleIf(val: string) {
-    this.templateQuestion.visibleIf = val;
-  }
-  public get enableIf(): string {
-    return this.templateQuestion.enableIf;
-  }
-  public set enableIf(val: string) {
-    this.templateQuestion.enableIf = val;
-  }
-  public get requiredIf(): string {
-    return this.templateQuestion.requiredIf;
-  }
-  public set requiredIf(val: string) {
-    this.templateQuestion.requiredIf = val;
-  }
-  public get isUnique(): boolean {
-    return this.getPropertyValue("isUnique");
-  }
-  public set isUnique(val: boolean) {
-    this.setPropertyValue("isUnique", val);
-  }
-  public get showInMultipleColumns(): boolean {
-    return this.getPropertyValue("showInMultipleColumns", false);
-  }
-  public set showInMultipleColumns(val: boolean) {
-    this.setPropertyValue("showInMultipleColumns", val);
-  }
-  public get isSupportMultipleColumns(): boolean {
-    return ["checkbox", "radiogroup"].indexOf(this.cellType) > -1;
-  }
-  public get isShowInMultipleColumns(): boolean {
-    return this.showInMultipleColumns && this.isSupportMultipleColumns;
-  }
-  public get validators(): Array<SurveyValidator> {
-    return this.templateQuestion.validators;
-  }
-  public set validators(val: Array<SurveyValidator>) {
-    this.templateQuestion.validators = val;
-  }
-  public get totalType(): string {
-    return this.getPropertyValue("totalType");
-  }
-  public set totalType(val: string) {
-    this.setPropertyValue("totalType", val);
-  }
-  public get totalExpression(): string {
-    return this.getPropertyValue("totalExpression");
-  }
-  public set totalExpression(val: string) {
-    this.setPropertyValue("totalExpression", val);
-  }
-  public get hasTotal(): boolean {
-    return this.totalType != "none" || !!this.totalExpression;
-  }
-  public get totalFormat(): string {
-    return this.getLocalizableStringText("totalFormat", "");
-  }
-  public set totalFormat(val: string) {
-    this.setLocalizableStringText("totalFormat", val);
-  }
-  get locTotalFormat(): LocalizableString {
-    return this.getLocalizableString("totalFormat");
-  }
-  public get renderAs(): string {
-    return this.getPropertyValue("renderAs");
-  }
-  public set renderAs(val: string) {
-    this.setPropertyValue("renderAs", val);
-  }
-  public get totalMaximumFractionDigits(): number {
-    return this.getPropertyValue("totalMaximumFractionDigits");
-  }
-  public set totalMaximumFractionDigits(val: number) {
-    if (val < -1 || val > 20) return;
-    this.setPropertyValue("totalMaximumFractionDigits", val);
-  }
-  public get totalMinimumFractionDigits(): number {
-    return this.getPropertyValue("totalMinimumFractionDigits");
-  }
-  public set totalMinimumFractionDigits(val: number) {
-    if (val < -1 || val > 20) return;
-    this.setPropertyValue("totalMinimumFractionDigits", val);
-  }
-  public get totalDisplayStyle(): string {
-    return this.getPropertyValue("totalDisplayStyle");
-  }
-  public set totalDisplayStyle(val: string) {
-    this.setPropertyValue("totalDisplayStyle", val);
-  }
-  public get totalCurrency(): string {
-    return this.getPropertyValue("totalCurrency");
-  }
-  public set totalCurrency(val: string) {
-    if (getCurrecyCodes().indexOf(val) < 0) return;
-    this.setPropertyValue("totalCurrency", val);
-  }
-  public get minWidth(): string {
-    return this.getPropertyValue("minWidth", "");
-  }
-  public set minWidth(val: string) {
-    this.setPropertyValue("minWidth", val);
-  }
-  public get width(): string {
-    return this.getPropertyValue("width", "");
-  }
-  public set width(val: string) {
-    this.setPropertyValue("width", val);
-  }
-  public get colCount(): number {
-    return this.getPropertyValue("colCount");
-  }
-  public set colCount(val: number) {
-    if (val < -1 || val > 4) return;
-    this.setPropertyValue("colCount", val);
-  }
-  public getLocale(): string {
-    return this.colOwner ? this.colOwner.getLocale() : "";
-  }
-  public getMarkdownHtml(text: string, name: string): string {
-    return this.colOwner ? this.colOwner.getMarkdownHtml(text, name) : null;
-  }
-  public getRenderer(name: string): string {
-    return !!this.colOwner ? this.colOwner.getRenderer(name) : null;
-  }
-  public getProcessedText(text: string): string {
-    return this.colOwner ? this.colOwner.getProcessedText(text) : text;
-  }
-  public createCellQuestion(row: MatrixDropdownRowModelBase): Question {
-    var qType = this.calcCellQuestionType(row);
-    var cellQuestion = <Question>this.createNewQuestion(qType);
-    this.callOnCellQuestionUpdate(cellQuestion, row);
-    return cellQuestion;
-  }
-  public updateCellQuestion(
-    cellQuestion: Question,
-    data: any,
-    onUpdateJson: (json: any) => any = null
-  ) {
-    this.setQuestionProperties(cellQuestion, onUpdateJson);
-    this.callOnCellQuestionUpdate(cellQuestion, data);
-  }
-  private callOnCellQuestionUpdate(cellQuestion: Question, data: any) {
-    var qType = cellQuestion.getType();
-    var qDefinition = (<any>matrixDropdownColumnTypes)[qType];
-    if (qDefinition && qDefinition["onCellQuestionUpdate"]) {
-      qDefinition["onCellQuestionUpdate"](
-        cellQuestion,
-        this,
-        this.colOwner,
-        data
-      );
-    }
-  }
-  defaultCellTypeChanged() {
-    this.updateTemplateQuestion();
-  }
-  protected calcCellQuestionType(row: MatrixDropdownRowModelBase): string {
-    let cellType = this.getDefaultCellQuestionType();
-    if(!!row && !!this.colOwner) {
-      cellType = this.colOwner.getCustomCellType(this, row, cellType);
-    }
-    return cellType;
-  }
-  private getDefaultCellQuestionType(): string {
-    if (this.cellType !== "default") return this.cellType;
-    if (this.colOwner) return this.colOwner.getCellType();
-    return settings.matrixDefaultCellType;
-  }
-  protected updateTemplateQuestion() {
-    var prevCellType = this.templateQuestion
-      ? this.templateQuestion.getType()
-      : "";
-    var curCellType = this.calcCellQuestionType(null);
-    if (curCellType === prevCellType) return;
-    if (this.templateQuestion) {
-      this.removeProperties(prevCellType);
-    }
-    this.templateQuestionValue = this.createNewQuestion(curCellType);
-    this.templateQuestion.locOwner = this;
-    this.addProperties(curCellType);
-    this.templateQuestion.onPropertyChanged.add((sender, options) => {
-      this.propertyValueChanged(
-        options.name,
-        options.oldValue,
-        options.newValue
-      );
-    });
-    this.templateQuestion.onItemValuePropertyChanged.add((sender, options) => {
-      this.doItemValuePropertyChanged(
-        options.propertyName,
-        options.obj,
-        options.name,
-        options.newValue,
-        options.oldValue
-      );
-    });
-    this.templateQuestion.isContentElement = true;
-    if (!this.isLoadingFromJson) {
-      this.templateQuestion.onGetSurvey = () => {
-        return this.getSurvey();
-      };
-    }
-    this.templateQuestion.locTitle.strChanged();
-  }
-  protected createNewQuestion(cellType: string): Question {
-    var question = <Question>Serializer.createClass(cellType);
-    if (!question) {
-      question = <Question>Serializer.createClass("text");
-    }
-    question.loadingOwner = this;
-    question.isEditableTemplateElement = true;
-    this.setQuestionProperties(question);
-    return question;
-  }
-  protected setQuestionProperties(
-    question: Question,
-    onUpdateJson: (json: any) => any = null
-  ) {
-    if (this.templateQuestion) {
-      var json = new JsonObject().toJsonObject(this.templateQuestion, true);
-      if (onUpdateJson) {
-        onUpdateJson(json);
-      }
-      json.type = question.getType();
-      new JsonObject().toObject(json, question);
-    }
-  }
-  protected propertyValueChanged(name: string, oldValue: any, newValue: any) {
-    super.propertyValueChanged(name, oldValue, newValue);
-    if (!Serializer.hasOriginalProperty(this, name)) return;
-    if (this.colOwner != null && !this.isLoadingFromJson) {
-      this.colOwner.onColumnPropertyChanged(this, name, newValue);
-    }
-  }
-  private doItemValuePropertyChanged(
-    propertyName: string,
-    obj: ItemValue,
-    name: string,
-    newValue: any,
-    oldValue: any
-  ) {
-    if (!Serializer.hasOriginalProperty(obj, name)) return;
-    if (this.colOwner != null && !this.isLoadingFromJson) {
-      this.colOwner.onColumnItemValuePropertyChanged(
-        this,
-        propertyName,
-        obj,
-        name,
-        newValue,
-        oldValue
-      );
-    }
-  }
-
-  private doShowInMultipleColumnsChanged() {
-    if (this.colOwner != null && !this.isLoadingFromJson) {
-      this.colOwner.onShowInMultipleColumnsChanged(this);
-    }
-  }
-  private getProperties(curCellType: string): Array<JsonObjectProperty> {
-    return Serializer.getDynamicPropertiesByObj(this, curCellType);
-  }
-  private removeProperties(curCellType: string) {
-    var properties = this.getProperties(curCellType);
-    for (var i = 0; i < properties.length; i++) {
-      var prop = properties[i];
-      delete (<any>this)[prop.name];
-      if (prop.serializationProperty) {
-        delete (<any>this)[prop.serializationProperty];
-      }
-    }
-  }
-  private addProperties(curCellType: string) {
-    var question = this.templateQuestion;
-    var properties = this.getProperties(curCellType);
-    for (var i = 0; i < properties.length; i++) {
-      var prop = properties[i];
-      this.addProperty(question, prop.name, false);
-      if (prop.serializationProperty) {
-        this.addProperty(question, prop.serializationProperty, true);
-      }
-    }
-  }
-  private addProperty(
-    question: Question,
-    propName: string,
-    isReadOnly: boolean
-  ) {
-    var desc = {
-      configurable: true,
-      get: function() {
-        return (<any>question)[propName];
-      },
-    };
-    if (!isReadOnly) {
-      (<any>desc)["set"] = function(v: any) {
-        (<any>question)[propName] = v;
-      };
-    }
-    Object.defineProperty(this, propName, desc);
-  }
-}
-
 export class MatrixDropdownCell {
   private questionValue: Question;
   constructor(
@@ -721,7 +86,7 @@ export class MatrixDropdownCell {
     data: IMatrixDropdownData
   ): Question {
     var res = data.createQuestion(this.row, this.column);
-    res.validateValueCallback = function() {
+    res.validateValueCallback = function () {
       return data.validateCell(row, column.name, row.value);
     };
     CustomPropertiesCollection.getProperties(column.getType()).forEach(
@@ -772,7 +137,7 @@ export class MatrixDropdownTotalCell extends MatrixDropdownCell {
   }
   public updateCellQuestion() {
     this.question.locCalculation();
-    this.column.updateCellQuestion(this.question, null, function(json) {
+    this.column.updateCellQuestion(this.question, null, function (json) {
       delete json["defaultValue"];
     });
     this.question.expression = this.getTotalExpression();
@@ -1026,7 +391,7 @@ implements ISurveyData, ISurveyImpl, ILocalizableOwner {
   getVariable(name: string): any {
     return undefined;
   }
-  setVariable(name: string, newValue: any) {}
+  setVariable(name: string, newValue: any) { }
   public getComment(name: string): string {
     var question = this.getQuestionByName(name);
     return !!question ? question.comment : "";
@@ -1234,7 +599,7 @@ implements ISurveyData, ISurveyImpl, ILocalizableOwner {
     for (var i = 0; i < questions.length; i++) {
       questions[i].readOnly = parentIsReadOnly;
     }
-    if(!!this.detailPanel) {
+    if (!!this.detailPanel) {
       this.detailPanel.readOnly = parentIsReadOnly;
     }
   }
@@ -1397,800 +762,10 @@ export class MatrixDropdownTotalRowModel extends MatrixDropdownRowModelBase {
   }
 }
 
-export class QuestionMatrixDropdownRenderedCell {
-  private static counter = 1;
-  private idValue: number;
-  private itemValue: ItemValue;
-  public minWidth: string = "";
-  public width: string = "";
-  public locTitle: LocalizableString;
-  public cell: MatrixDropdownCell;
-  public column: MatrixDropdownColumn;
-  public row: MatrixDropdownRowModelBase;
-  public question: Question;
-  public isRemoveRow: boolean;
-  public choiceIndex: number;
-  public matrix: QuestionMatrixDropdownModelBase;
-  public requiredText: string;
-  public isEmpty: boolean;
-  public colSpans: number = 1;
-  public panel: PanelModel;
-  public isShowHideDetail: boolean;
-  public isActionsCell: boolean = false;
-  public className: string = "";
-  public isDragHandlerCell: boolean = false;
-  public constructor() {
-    this.idValue = QuestionMatrixDropdownRenderedCell.counter++;
-  }
-  public get hasQuestion(): boolean {
-    return !!this.question;
-  }
-  public get hasTitle(): boolean {
-    return !!this.locTitle;
-  }
-  public get hasPanel(): boolean {
-    return !!this.panel;
-  }
-  public get id(): number {
-    return this.idValue;
-  }
-  public get showErrorOnTop(): boolean {
-    return this.showErrorOnCore("top");
-  }
-  public get showErrorOnBottom(): boolean {
-    return this.showErrorOnCore("bottom");
-  }
-  private showErrorOnCore(location: string): boolean {
-    return (
-      this.getShowErrorLocation() == location &&
-      (!this.isChoice || this.isFirstChoice)
-    );
-  }
-  private getShowErrorLocation(): string {
-    return this.hasQuestion ? this.question.survey.questionErrorLocation : "";
-  }
-  public get item(): ItemValue {
-    return this.itemValue;
-  }
-  public set item(val: ItemValue) {
-    this.itemValue = val;
-    if (!!val) {
-      val.hideCaption = true;
-    }
-  }
-  public get isChoice(): boolean {
-    return !!this.item;
-  }
-  public get choiceValue(): any {
-    return this.isChoice ? this.item.value : null;
-  }
-  public get isCheckbox(): boolean {
-    return this.isChoice && this.question.getType() == "checkbox";
-  }
-  public get isFirstChoice(): boolean {
-    return this.choiceIndex === 0;
-  }
-  public get css(): string {
-    return new CssClassBuilder()
-      .append(this.className)
-      .append(this.question.cssError, this.question.errors.length > 0)
-      .toString();
-  }
-  public get headers(): string {
-    if (
-      this.cell &&
-      this.cell.column &&
-      this.cell.column.isShowInMultipleColumns
-    ) {
-      return this.item.locText.renderedHtml;
-    }
-    if (this.question && this.question.isVisible) {
-      return this.question.locTitle.renderedHtml;
-    }
-    if (this.hasTitle) {
-      return this.locTitle.renderedHtml || "";
-    }
-    return "";
-  }
-
-  public calculateFinalClassName(matrixCssClasses: any): string {
-    const questionCss = this.cell.question.cssClasses;
-    // 'text-align': $data.isChoice ? 'center':
-    const builder = new CssClassBuilder()
-      .append(questionCss.itemValue, !!questionCss)
-      .append(questionCss.asCell, !!questionCss);
-    return builder.append(matrixCssClasses.cell, builder.isEmpty() && !!matrixCssClasses)
-      .append(questionCss.hasError, this.question.errors.length > 0)
-      .append(matrixCssClasses.choiceCell, this.isChoice)
-      .toString();
-  }
-}
-
-export class QuestionMatrixDropdownRenderedRow extends Base {
-  @property({ defaultValue: null }) ghostPosition: string;
-
-  public isDetailRow: boolean = false;
-  public row: MatrixDropdownRowModelBase;
-  private static counter = 1;
-  private idValue: number;
-  public cells: Array<QuestionMatrixDropdownRenderedCell> = [];
-  public className: string = "";
-
-  public constructor(public cssClasses: any) {
-    super();
-    this.onCreating();
-    this.idValue = QuestionMatrixDropdownRenderedRow.counter++;
-  }
-  public onCreating() {} // need for knockout binding see QuestionMatrixDropdownRenderedRow.prototype["onCreating"]
-  public get id(): number {
-    return this.idValue;
-  }
-  public get attributes() {
-    if (!this.row) return {};
-    return { "data-sv-drop-target-matrix-row": this.row.id };
-  }
-  public get ghostPositionCssClass() {
-    if (this.ghostPosition === "top")
-      return this.cssClasses.dragDropGhostPositionTop;
-    if (this.ghostPosition === "bottom")
-      return this.cssClasses.dragDropGhostPositionBottom;
-    return "";
-  }
-}
-
-export class QuestionMatrixDropdownRenderedTable extends Base {
-  private headerRowValue: QuestionMatrixDropdownRenderedRow;
-  private footerRowValue: QuestionMatrixDropdownRenderedRow;
-  private hasRemoveRowsValue: boolean;
-  private rowsActions: Array<Array<IAction>>;
-  private cssClasses: any;
-  public constructor(public matrix: QuestionMatrixDropdownModelBase) {
-    super();
-    this.createNewArray("rows");
-    this.build();
-  }
-  public get showTable(): boolean {
-    return this.getPropertyValue("showTable", true);
-  }
-  public get showHeader(): boolean {
-    return this.getPropertyValue("showHeader");
-  }
-  public get showAddRowOnTop(): boolean {
-    return this.getPropertyValue("showAddRowOnTop", false);
-  }
-  public get showAddRowOnBottom(): boolean {
-    return this.getPropertyValue("showAddRowOnBottom", false);
-  }
-  public get showFooter(): boolean {
-    return this.matrix.hasFooter && this.matrix.isColumnLayoutHorizontal;
-  }
-  public get hasFooter(): boolean {
-    return !!this.footerRow;
-  }
-  public get hasRemoveRows(): boolean {
-    return this.hasRemoveRowsValue;
-  }
-  public isRequireReset(): boolean {
-    return (
-      this.hasRemoveRows != this.matrix.canRemoveRows ||
-      !this.matrix.isColumnLayoutHorizontal
-    );
-  }
-  public get headerRow(): QuestionMatrixDropdownRenderedRow {
-    return this.headerRowValue;
-  }
-  public get footerRow(): QuestionMatrixDropdownRenderedRow {
-    return this.footerRowValue;
-  }
-  public get rows(): Array<QuestionMatrixDropdownRenderedRow> {
-    return this.getPropertyValue("rows");
-  }
-  protected build() {
-    this.hasRemoveRowsValue = this.matrix.canRemoveRows;
-    //build rows now
-    var rows = this.matrix.visibleRows;
-    this.cssClasses = this.matrix.cssClasses;
-    this.buildRowsActions();
-    this.buildHeader();
-    this.buildRows();
-    this.buildFooter();
-    this.updateShowTableAndAddRow();
-  }
-  public updateShowTableAndAddRow() {
-    var showTable =
-      this.rows.length > 0 ||
-      this.matrix.isDesignMode ||
-      !this.matrix.getShowColumnsIfEmpty();
-    this.setPropertyValue("showTable", showTable);
-    var showAddRow = this.matrix.canAddRow && showTable;
-    var showAddRowOnTop = showAddRow;
-    var showAddRowOnBottom = showAddRow;
-    if (showAddRowOnTop) {
-      if (this.matrix.getAddRowLocation() === "default") {
-        showAddRowOnTop = this.matrix.columnLayout === "vertical";
-      } else {
-        showAddRowOnTop = this.matrix.getAddRowLocation() !== "bottom";
-      }
-    }
-    if (showAddRowOnBottom && this.matrix.getAddRowLocation() !== "topBottom") {
-      showAddRowOnBottom = !showAddRowOnTop;
-    }
-    this.setPropertyValue("showAddRowOnTop", showAddRowOnTop);
-    this.setPropertyValue("showAddRowOnBottom", showAddRowOnBottom);
-  }
-  public onAddedRow() {
-    if (this.getRenderedDataRowCount() >= this.matrix.visibleRows.length)
-      return;
-    var row = this.matrix.visibleRows[this.matrix.visibleRows.length - 1];
-    this.rowsActions.push(this.buildRowActions(row));
-    this.addHorizontalRow(
-      this.rows,
-      row,
-      this.matrix.visibleRows.length == 1 && !this.matrix.showHeader
-    );
-    this.updateShowTableAndAddRow();
-  }
-  private getRenderedDataRowCount(): number {
-    var res = 0;
-    for (var i = 0; i < this.rows.length; i++) {
-      if (!this.rows[i].isDetailRow) res++;
-    }
-    return res;
-  }
-  public onRemovedRow(row: MatrixDropdownRowModelBase) {
-    var rowIndex = this.getRenderedRowIndex(row);
-    if (rowIndex < 0) return;
-    this.rowsActions.splice(rowIndex, 1);
-    var removeCount = 1;
-    if (
-      rowIndex < this.rows.length - 1 &&
-      this.rows[rowIndex + 1].isDetailRow
-    ) {
-      removeCount++;
-    }
-    this.rows.splice(rowIndex, removeCount);
-    this.updateShowTableAndAddRow();
-  }
-  public onDetailPanelChangeVisibility(
-    row: MatrixDropdownRowModelBase,
-    isShowing: boolean
-  ) {
-    var rowIndex = this.getRenderedRowIndex(row);
-    if (rowIndex < 0) return;
-    var panelRowIndex =
-      rowIndex < this.rows.length - 1 && this.rows[rowIndex + 1].isDetailRow
-        ? rowIndex + 1
-        : -1;
-    if ((isShowing && panelRowIndex > -1) || (!isShowing && panelRowIndex < 0))
-      return;
-    if (isShowing) {
-      var detailRow = this.createDetailPanelRow(row, this.rows[rowIndex]);
-      this.rows.splice(rowIndex + 1, 0, detailRow);
-    } else {
-      this.rows.splice(panelRowIndex, 1);
-    }
-  }
-  private getRenderedRowIndex(row: MatrixDropdownRowModelBase): number {
-    for (var i = 0; i < this.rows.length; i++) {
-      if (this.rows[i].row == row) return i;
-    }
-    return -1;
-  }
-  protected buildRowsActions() {
-    this.rowsActions = [];
-    var rows = this.matrix.visibleRows;
-    for (var i = 0; i < rows.length; i++) {
-      this.rowsActions.push(this.buildRowActions(rows[i]));
-    }
-  }
-  protected buildHeader() {
-    var colHeaders =
-      this.matrix.isColumnLayoutHorizontal && this.matrix.showHeader;
-    var isShown =
-      colHeaders ||
-      (this.matrix.hasRowText && !this.matrix.isColumnLayoutHorizontal);
-    this.setPropertyValue("showHeader", isShown);
-    if (!isShown) return;
-    this.headerRowValue = new QuestionMatrixDropdownRenderedRow(
-      this.cssClasses
-    );
-    if(this.matrix.allowRowsDragAndDrop) {
-      this.headerRow.cells.push(this.createHeaderCell(null));
-    }
-    if (this.hasActionCellInRows("start")) {
-      this.headerRow.cells.push(this.createHeaderCell(null));
-    }
-    if (this.matrix.hasRowText && this.matrix.showHeader) {
-      this.headerRow.cells.push(this.createHeaderCell(null));
-    }
-    if (this.matrix.isColumnLayoutHorizontal) {
-      for (var i = 0; i < this.matrix.visibleColumns.length; i++) {
-        var column = this.matrix.visibleColumns[i];
-        if (!column.hasVisibleCell) continue;
-        if (column.isShowInMultipleColumns) {
-          this.createMutlipleColumnsHeader(column);
-        } else {
-          this.headerRow.cells.push(this.createHeaderCell(column));
-        }
-      }
-    } else {
-      var rows = this.matrix.visibleRows;
-      for (var i = 0; i < rows.length; i++) {
-        this.headerRow.cells.push(this.createTextCell(rows[i].locText));
-      }
-      if (this.matrix.hasFooter) {
-        this.headerRow.cells.push(
-          this.createTextCell(this.matrix.getFooterText())
-        );
-      }
-    }
-    if (this.hasActionCellInRows("end")) {
-      this.headerRow.cells.push(this.createHeaderCell(null));
-    }
-  }
-  protected buildFooter() {
-    if (!this.showFooter) return;
-    this.footerRowValue = new QuestionMatrixDropdownRenderedRow(
-      this.cssClasses
-    );
-    if(this.matrix.allowRowsDragAndDrop) {
-      this.footerRow.cells.push(this.createHeaderCell(null));
-    }
-    if (this.hasActionCellInRows("start")) {
-      this.footerRow.cells.push(this.createHeaderCell(null));
-    }
-    if (this.matrix.hasRowText) {
-      this.footerRow.cells.push(
-        this.createTextCell(this.matrix.getFooterText())
-      );
-    }
-    var cells = this.matrix.visibleTotalRow.cells;
-    for (var i = 0; i < cells.length; i++) {
-      var cell = cells[i];
-      if (!cell.column.hasVisibleCell) continue;
-      if (cell.column.isShowInMultipleColumns) {
-        this.createMutlipleColumnsFooter(this.footerRow, cell);
-      } else {
-        this.footerRow.cells.push(this.createEditCell(cell));
-      }
-    }
-    if (this.hasActionCellInRows("end")) {
-      this.footerRow.cells.push(this.createHeaderCell(null));
-    }
-  }
-  protected buildRows() {
-    var rows = this.matrix.isColumnLayoutHorizontal
-      ? this.buildHorizontalRows()
-      : this.buildVerticalRows();
-    this.setPropertyValue("rows", rows);
-  }
-  private hasActionCellInRowsValues: any = {};
-  private hasActionCellInRows(location: "start" | "end"): boolean {
-    if (this.hasActionCellInRowsValues[location] === undefined) {
-      var rows = this.matrix.visibleRows;
-      this.hasActionCellInRowsValues[location] = false;
-      for (var i = 0; i < rows.length; i++) {
-        if (!this.isValueEmpty(this.getRowActions(i, location))) {
-          this.hasActionCellInRowsValues[location] = true;
-          break;
-        }
-      }
-    }
-    return this.hasActionCellInRowsValues[location];
-  }
-  private canRemoveRow(row: MatrixDropdownRowModelBase): boolean {
-    return this.matrix.canRemoveRow(row);
-  }
-  private buildHorizontalRows(): Array<QuestionMatrixDropdownRenderedRow> {
-    var rows = this.matrix.visibleRows;
-    var renderedRows: Array<QuestionMatrixDropdownRenderedRow> = [];
-    for (var i = 0; i < rows.length; i++) {
-      this.addHorizontalRow(
-        renderedRows,
-        rows[i],
-        i == 0 && !this.matrix.showHeader
-      );
-    }
-    return renderedRows;
-  }
-  private addHorizontalRow(
-    renderedRows: Array<QuestionMatrixDropdownRenderedRow>,
-    row: MatrixDropdownRowModelBase,
-    useAsHeader: boolean
-  ) {
-    var renderedRow = this.createHorizontalRow(row, useAsHeader);
-    renderedRow.row = row;
-    renderedRows.push(renderedRow);
-    if (row.isDetailPanelShowing) {
-      renderedRows.push(this.createDetailPanelRow(row, renderedRow));
-    }
-  }
-  private getRowDragCell(rowIndex: number) {
-    const cell = new QuestionMatrixDropdownRenderedCell();
-    cell.isDragHandlerCell = true;
-    cell.className = this.cssClasses.actionsCell;
-    cell.row = this.matrix.visibleRows[rowIndex];
-    return cell;
-  }
-  private getRowActionsCell(rowIndex: number, location: "start" | "end") {
-    const rowActions = this.getRowActions(rowIndex, location);
-    if (!this.isValueEmpty(rowActions)) {
-      const cell = new QuestionMatrixDropdownRenderedCell();
-      const actionContainer = new AdaptiveActionContainer();
-      actionContainer.setItems(rowActions);
-
-      const itemValue = new ItemValue(actionContainer);
-      cell.item = itemValue;
-      cell.isActionsCell = true;
-      cell.className = this.cssClasses.actionsCell;
-      cell.row = this.matrix.visibleRows[rowIndex];
-      return cell;
-    }
-    return null;
-  }
-  private getRowActions(rowIndex: number, location: "start" | "end") {
-    var actions = this.rowsActions[rowIndex];
-    if (!Array.isArray(actions)) return [];
-    return actions.filter((action) => {
-      if (!action.location) {
-        action.location = "start";
-      }
-      return action.location === location;
-    });
-  }
-  private buildRowActions(row: MatrixDropdownRowModelBase): Array<IAction> {
-    var actions: Array<IAction> = [];
-    this.setDefaultRowActions(row, actions);
-    if (!!this.matrix.survey) {
-      actions = this.matrix.survey.getUpdatedMatrixRowActions(
-        this.matrix,
-        row,
-        actions
-      );
-    }
-    return actions;
-  }
-  protected setDefaultRowActions(
-    row: MatrixDropdownRowModelBase,
-    actions: Array<IAction>
-  ) {
-    if (this.hasRemoveRows && this.canRemoveRow(row)) {
-      actions.push(
-        new Action({
-          id: "remove-row",
-          location: "end",
-          enabled: !this.matrix.isInputReadOnly,
-          component: "sv-matrix-remove-button",
-          data: { row: row, question: this.matrix },
-        })
-      );
-    }
-
-    if (row.hasPanel) {
-      actions.push(
-        new Action({
-          id: "show-detail",
-          title: surveyLocalization.getString("editText"),
-          showTitle: false,
-          location: "start",
-          component: "sv-matrix-detail-button",
-          data: { row: row, question: this.matrix },
-        })
-      );
-    }
-  }
-  private createHorizontalRow(
-    row: MatrixDropdownRowModelBase,
-    useAsHeader: boolean
-  ): QuestionMatrixDropdownRenderedRow {
-    var res = new QuestionMatrixDropdownRenderedRow(this.cssClasses);
-    if(this.matrix.allowRowsDragAndDrop) {
-      var rowIndex = this.matrix.visibleRows.indexOf(row);
-      res.cells.push(this.getRowDragCell(rowIndex));
-    }
-    this.addRowActionsCell(row, res, "start");
-    if (this.matrix.hasRowText) {
-      var renderedCell = this.createTextCell(row.locText);
-      renderedCell.row = row;
-      res.cells.push(renderedCell);
-      if (useAsHeader) {
-        this.setHeaderCellWidth(null, renderedCell);
-      }
-      renderedCell.className = new CssClassBuilder()
-        .append(renderedCell.className)
-        .append(this.cssClasses.detailRowText, row.hasPanel)
-        .toString();
-    }
-    for (var i = 0; i < row.cells.length; i++) {
-      let cell = row.cells[i];
-      if (!cell.column.hasVisibleCell) continue;
-      if (cell.column.isShowInMultipleColumns) {
-        this.createMutlipleEditCells(res, cell);
-      } else {
-        var renderedCell = this.createEditCell(cell);
-        res.cells.push(renderedCell);
-        if (useAsHeader) {
-          this.setHeaderCellWidth(cell.column, renderedCell);
-        }
-      }
-    }
-    this.addRowActionsCell(row, res, "end");
-    return res;
-  }
-  private addRowActionsCell(
-    row: MatrixDropdownRowModelBase,
-    renderedRow: QuestionMatrixDropdownRenderedRow,
-    location: "start" | "end"
-  ) {
-    var rowIndex = this.matrix.visibleRows.indexOf(row);
-    if (this.hasActionCellInRows(location)) {
-      const actions = this.getRowActionsCell(rowIndex, location);
-      if (!!actions) {
-        renderedRow.cells.push(actions);
-      } else {
-        var cell = new QuestionMatrixDropdownRenderedCell();
-        cell.isEmpty = true;
-        renderedRow.cells.push(cell);
-      }
-    }
-  }
-  private createDetailPanelRow(
-    row: MatrixDropdownRowModelBase,
-    renderedRow: QuestionMatrixDropdownRenderedRow
-  ): QuestionMatrixDropdownRenderedRow {
-    var res = new QuestionMatrixDropdownRenderedRow(this.cssClasses);
-    res.row = row;
-    res.className += this.cssClasses.detailRow;
-    res.isDetailRow = true;
-    var buttonCell = new QuestionMatrixDropdownRenderedCell();
-    if (this.matrix.hasRowText) {
-      buttonCell.colSpans = 2;
-    }
-    buttonCell.isEmpty = true;
-    res.cells.push(buttonCell);
-    var actionsCell = null;
-    if (this.hasActionCellInRows("end")) {
-      actionsCell = new QuestionMatrixDropdownRenderedCell();
-      actionsCell.isEmpty = true;
-    }
-    var cell = new QuestionMatrixDropdownRenderedCell();
-    cell.panel = row.detailPanel;
-    cell.colSpans =
-      renderedRow.cells.length -
-      buttonCell.colSpans -
-      (!!actionsCell ? actionsCell.colSpans : 0);
-    cell.className = this.cssClasses.detailPanelCell;
-    res.cells.push(cell);
-    if (!!actionsCell) {
-      res.cells.push(actionsCell);
-    }
-    if (
-      typeof this.matrix.onCreateDetailPanelRenderedRowCallback === "function"
-    ) {
-      this.matrix.onCreateDetailPanelRenderedRowCallback(res);
-    }
-    return res;
-  }
-
-  private buildVerticalRows(): Array<QuestionMatrixDropdownRenderedRow> {
-    var columns = this.matrix.columns;
-    var renderedRows = [];
-    for (var i = 0; i < columns.length; i++) {
-      var col = columns[i];
-      if (col.isVisible && col.hasVisibleCell) {
-        if (col.isShowInMultipleColumns) {
-          this.createMutlipleVerticalRows(renderedRows, col, i);
-        } else {
-          renderedRows.push(this.createVerticalRow(col, i));
-        }
-      }
-    }
-    if (this.hasActionCellInRows("end")) {
-      renderedRows.push(this.createEndVerticalActionRow());
-    }
-    return renderedRows;
-  }
-  private createMutlipleVerticalRows(
-    renderedRows: Array<QuestionMatrixDropdownRenderedRow>,
-    column: MatrixDropdownColumn,
-    index: number
-  ) {
-    var choices = this.getMultipleColumnChoices(column);
-    if (!choices) return;
-    for (var i = 0; i < choices.length; i++) {
-      renderedRows.push(this.createVerticalRow(column, index, choices[i], i));
-    }
-  }
-  private createVerticalRow(
-    column: MatrixDropdownColumn,
-    index: number,
-    choice: ItemValue = null,
-    choiceIndex: number = -1
-  ): QuestionMatrixDropdownRenderedRow {
-    var res = new QuestionMatrixDropdownRenderedRow(this.cssClasses);
-    if (this.matrix.showHeader) {
-      var lTitle = !!choice ? choice.locText : column.locTitle;
-      var hCell = this.createTextCell(lTitle);
-      hCell.column = column;
-      if (!choice) {
-        this.setRequriedToHeaderCell(column, hCell);
-      }
-      res.cells.push(hCell);
-    }
-    var rows = this.matrix.visibleRows;
-    for (var i = 0; i < rows.length; i++) {
-      var rChoice = choice;
-      var rChoiceIndex = choiceIndex >= 0 ? choiceIndex : i;
-      var cell = rows[i].cells[index];
-      var visChoices = !!choice ? cell.question.visibleChoices : undefined;
-      if (!!visChoices && rChoiceIndex < visChoices.length) {
-        rChoice = visChoices[rChoiceIndex];
-      }
-      var rCell = this.createEditCell(cell, rChoice);
-      rCell.item = rChoice;
-      rCell.choiceIndex = rChoiceIndex;
-      res.cells.push(rCell);
-    }
-    if (this.matrix.hasTotal) {
-      res.cells.push(
-        this.createEditCell(this.matrix.visibleTotalRow.cells[index])
-      );
-    }
-
-    return res;
-  }
-  private createEndVerticalActionRow(): QuestionMatrixDropdownRenderedRow {
-    var res = new QuestionMatrixDropdownRenderedRow(this.cssClasses);
-    if (this.matrix.showHeader) {
-      res.cells.push(this.createTextCell(null));
-    }
-    var rows = this.matrix.visibleRows;
-    for (var i = 0; i < rows.length; i++) {
-      res.cells.push(this.getRowActionsCell(i, "end"));
-    }
-    if (this.matrix.hasTotal) {
-      res.cells.push(this.createTextCell(null));
-    }
-    return res;
-  }
-  private createMutlipleEditCells(
-    rRow: QuestionMatrixDropdownRenderedRow,
-    cell: MatrixDropdownCell,
-    isFooter: boolean = false
-  ) {
-    var choices = isFooter
-      ? this.getMultipleColumnChoices(cell.column)
-      : cell.question.visibleChoices;
-    if (!choices) return;
-    for (var i = 0; i < choices.length; i++) {
-      var rCell = this.createEditCell(cell, !isFooter ? choices[i] : undefined);
-      if (!isFooter) {
-        //rCell.item = choices[i];
-        rCell.choiceIndex = i;
-      }
-      rRow.cells.push(rCell);
-    }
-  }
-  private createEditCell(
-    cell: MatrixDropdownCell,
-    choiceItem: any = undefined
-  ): QuestionMatrixDropdownRenderedCell {
-    var res = new QuestionMatrixDropdownRenderedCell();
-    res.cell = cell;
-    res.row = cell.row;
-    res.question = cell.question;
-    res.matrix = this.matrix;
-    res.item = choiceItem;
-
-    res.className = res.calculateFinalClassName(this.cssClasses);
-    //res.css = res.calcCss(this.cssClasses.cell);
-
-    // var questionCss = cell.question.cssClasses;
-    // var className = "";
-    // if (!!questionCss) {
-    //   className = "";
-    //   if (!!questionCss.itemValue) {
-    //     className += " " + questionCss.itemValue;
-    //   }
-    //   if (!!questionCss.asCell) {
-    //     if (!!className) className += "";
-    //     className += questionCss.asCell;
-    //   }
-    // }
-    // if (!className && !!this.cssClasses.cell) {
-    //   className = this.cssClasses.cell;
-    // }
-    //res.className = className;
-    return res;
-  }
-  private createMutlipleColumnsFooter(
-    rRow: QuestionMatrixDropdownRenderedRow,
-    cell: MatrixDropdownCell
-  ) {
-    this.createMutlipleEditCells(rRow, cell, true);
-  }
-  private createMutlipleColumnsHeader(column: MatrixDropdownColumn) {
-    var choices = this.getMultipleColumnChoices(column);
-    if (!choices) return;
-    for (var i = 0; i < choices.length; i++) {
-      var cell = this.createTextCell(choices[i].locText);
-      this.setHeaderCell(column, cell);
-      this.headerRow.cells.push(cell);
-    }
-  }
-  private getMultipleColumnChoices(column: MatrixDropdownColumn): any {
-    var choices = column.templateQuestion.choices;
-    if (!!choices && Array.isArray(choices) && choices.length == 0)
-      return this.matrix.choices;
-    choices = column.templateQuestion.visibleChoices;
-    if (!choices || !Array.isArray(choices)) return null;
-    return choices;
-  }
-  private createHeaderCell(
-    column: MatrixDropdownColumn
-  ): QuestionMatrixDropdownRenderedCell {
-    var cell = this.createTextCell(!!column ? column.locTitle : null);
-    cell.column = column;
-    this.setHeaderCell(column, cell);
-    if (this.cssClasses.headerCell) {
-      cell.className = this.cssClasses.headerCell;
-    }
-    return cell;
-  }
-  private setHeaderCell(
-    column: MatrixDropdownColumn,
-    cell: QuestionMatrixDropdownRenderedCell
-  ) {
-    this.setHeaderCellWidth(column, cell);
-    this.setRequriedToHeaderCell(column, cell);
-  }
-  private setHeaderCellWidth(
-    column: MatrixDropdownColumn,
-    cell: QuestionMatrixDropdownRenderedCell
-  ) {
-    cell.minWidth = column != null ? this.matrix.getColumnWidth(column) : "";
-    cell.width = column != null ? column.width : this.matrix.getRowTitleWidth();
-  }
-  private setRequriedToHeaderCell(
-    column: MatrixDropdownColumn,
-    cell: QuestionMatrixDropdownRenderedCell
-  ) {
-    if (!!column && column.isRequired && this.matrix.survey) {
-      cell.requiredText = this.matrix.survey.requiredText;
-    }
-  }
-  private createRemoveRowCell(
-    row: MatrixDropdownRowModelBase
-  ): QuestionMatrixDropdownRenderedCell {
-    var res = new QuestionMatrixDropdownRenderedCell();
-    res.row = row;
-    res.isRemoveRow = this.canRemoveRow(row);
-    if (!!this.cssClasses.cell) {
-      res.className = this.cssClasses.cell;
-    }
-    return res;
-  }
-  private createTextCell(
-    locTitle: LocalizableString
-  ): QuestionMatrixDropdownRenderedCell {
-    var cell = new QuestionMatrixDropdownRenderedCell();
-    cell.locTitle = locTitle;
-    if (!!this.cssClasses.cell) {
-      cell.className = this.cssClasses.cell;
-    }
-    return cell;
-  }
-}
-
 /**
  * A base class for matrix dropdown and matrix dynamic questions.
  */
-export class QuestionMatrixDropdownModelBase
-  extends QuestionMatrixBaseModel<
-    MatrixDropdownRowModelBase,
-    MatrixDropdownColumn
-  >
-  implements IMatrixDropdownData {
+export class QuestionMatrixDropdownModelBase extends QuestionMatrixBaseModel<MatrixDropdownRowModelBase, MatrixDropdownColumn> implements IMatrixDropdownData {
   public static get defaultCellType() {
     return settings.matrixDefaultCellType;
   }
@@ -2502,6 +1077,18 @@ export class QuestionMatrixDropdownModelBase
   }
   public set horizontalScroll(val: boolean) {
     this.setPropertyValue("horizontalScroll", val);
+  }
+  /**
+   * The Matrix toolbar and inner panel toolbars get adaptive if the property is set to true.
+   */
+  public get allowAdaptiveActions(): boolean {
+    return this.getPropertyValue("allowAdaptiveActions");
+  }
+  public set allowAdaptiveActions(val: boolean) {
+    this.setPropertyValue("allowAdaptiveActions", val);
+    if(!!this.detailPanel) {
+      this.detailPanel.allowAdaptiveActions = val;
+    }
   }
   public getRequiredText(): string {
     return this.survey ? this.survey.requiredText : "";
@@ -3026,7 +1613,7 @@ export class QuestionMatrixDropdownModelBase
     return questions;
   }
 
-  protected onBeforeValueChanged(val: any) {}
+  protected onBeforeValueChanged(val: any) { }
   private onSetQuestionValue() {
     if (this.isRowChanging) return;
     this.onBeforeValueChanged(this.value);
@@ -3560,74 +2147,6 @@ export class QuestionMatrixDropdownModelBase
 }
 
 Serializer.addClass(
-  "matrixdropdowncolumn",
-  [
-    { name: "!name", isUnique: true },
-    { name: "title", serializationProperty: "locTitle" },
-    {
-      name: "cellType",
-      default: "default",
-      choices: () => {
-        var res = MatrixDropdownColumn.getColumnTypes();
-        res.splice(0, 0, "default");
-        return res;
-      },
-    },
-    { name: "colCount", default: -1, choices: [-1, 0, 1, 2, 3, 4] },
-    "isRequired:boolean",
-    "isUnique:boolean",
-    {
-      name: "requiredErrorText:text",
-      serializationProperty: "locRequiredErrorText",
-    },
-    "readOnly:boolean",
-    "minWidth",
-    "width",
-    "visibleIf:condition",
-    "enableIf:condition",
-    "requiredIf:condition",
-    {
-      name: "showInMultipleColumns:boolean",
-      dependsOn: "cellType",
-      visibleIf: function(obj: any) {
-        if (!obj) return false;
-        return obj.isSupportMultipleColumns;
-      },
-    },
-    {
-      name: "validators:validators",
-      baseClassName: "surveyvalidator",
-      classNamePart: "validator",
-    },
-    {
-      name: "totalType",
-      default: "none",
-      choices: ["none", "sum", "count", "min", "max", "avg"],
-    },
-    "totalExpression:expression",
-    { name: "totalFormat", serializationProperty: "locTotalFormat" },
-    {
-      name: "totalDisplayStyle",
-      default: "none",
-      choices: ["none", "decimal", "currency", "percent"],
-    },
-    {
-      name: "totalCurrency",
-      choices: () => {
-        return getCurrecyCodes();
-      },
-      default: "USD",
-    },
-    { name: "totalMaximumFractionDigits:number", default: -1 },
-    { name: "totalMinimumFractionDigits:number", default: -1 },
-    { name: "renderAs", default: "default", visible: false },
-  ],
-  function() {
-    return new MatrixDropdownColumn("");
-  }
-);
-
-Serializer.addClass(
   "matrixdropdownbase",
   [
     {
@@ -3668,8 +2187,9 @@ Serializer.addClass(
     },
     { name: "columnColCount", default: 0, choices: [0, 1, 2, 3, 4] },
     "columnMinWidth",
+    { name: "allowAdaptiveActions:boolean", default: true },
   ],
-  function() {
+  function () {
     return new QuestionMatrixDropdownModelBase("");
   },
   "matrixbase"
