+++ tests/ko/survey_kotests.ts
@@ -7,11 +7,8 @@ import { Question } from "../../src/question";
 import { QuestionMatrix } from "../../src/knockout/koquestion_matrix";
 import { QuestionMatrixDropdown } from "../../src/knockout/koquestion_matrixdropdown";
 import { QuestionPanelDynamic } from "../../src/knockout/koquestion_paneldynamic";
-import { MatrixDropdownColumn } from "../../src/question_matrixdropdownbase";
-import {
-  QuestionMultipleText,
-  MultipleTextItem,
-} from "../../src/knockout/koquestion_multipletext";
+import { MatrixDropdownColumn } from "../../src/question_matrixdropdowncolumn";
+import { QuestionMultipleText, MultipleTextItem } from "../../src/knockout/koquestion_multipletext";
 import { Page, Panel, QuestionRow } from "../../src/knockout/kopage";
 import { CustomWidgetCollection } from "../../src/questionCustomWidgets";
 import { koTemplate } from "../../src/knockout/templateText";
@@ -29,7 +26,7 @@ import { ItemValue } from "../../src/itemvalue";
 
 export default QUnit.module("koTests");
 
-QUnit.test("Survey.koCurrentPage", function(assert) {
+QUnit.test("Survey.koCurrentPage", function (assert) {
   var survey = new Survey();
   survey.addPage(createPageWithQuestion("Page 1"));
   survey.addPage(createPageWithQuestion("Page 2"));
@@ -59,7 +56,7 @@ QUnit.test("Survey.koCurrentPage", function(assert) {
   assert.equal(survey.koIsFirstPage(), false, "is last page");
   assert.equal(survey.koIsLastPage(), true, "is last page");
 });
-QUnit.test("koOtherVisible for one choice items", function(assert) {
+QUnit.test("koOtherVisible for one choice items", function (assert) {
   var survey = new Survey();
   var question = new QuestionCheckbox("q");
   question.hasOther = true;
@@ -72,7 +69,7 @@ QUnit.test("koOtherVisible for one choice items", function(assert) {
     "Other visible is true after selecting it"
   );
 });
-QUnit.test("Create koValue as observable array for checkbox", function(assert) {
+QUnit.test("Create koValue as observable array for checkbox", function (assert) {
   var question = new QuestionCheckbox("q");
   question.koValue.push("test1");
   question.koValue.push("test2");
@@ -87,7 +84,7 @@ QUnit.test("Create koValue as observable array for checkbox", function(assert) {
     "value is set correctly."
   );
 });
-QUnit.test("Default value for checkbox", function(assert) {
+QUnit.test("Default value for checkbox", function (assert) {
   var survey = new Survey();
   survey.addNewPage("p1");
   var question = new QuestionCheckbox("q");
@@ -98,7 +95,7 @@ QUnit.test("Default value for checkbox", function(assert) {
     "the koValue by default should be empty array"
   );
 });
-QUnit.test("koOtherVisible for multi choice items", function(assert) {
+QUnit.test("koOtherVisible for multi choice items", function (assert) {
   var survey = new Survey();
   var question = new QuestionCheckbox("q");
   question.hasOther = true;
@@ -120,7 +117,7 @@ QUnit.test("koOtherVisible for multi choice items", function(assert) {
 });
 QUnit.test(
   "Update koValue on changing data in Survey or Question.value ",
-  function(assert) {
+  function (assert) {
     var survey = new Survey();
     survey.setValue("textQuestion", "initialValue");
     var page = survey.addNewPage("my page");
@@ -134,7 +131,7 @@ QUnit.test(
 );
 QUnit.test(
   "Update koValue on changing data in Survey or Question.value for Multiple Answer Question ",
-  function(assert) {
+  function (assert) {
     var survey = new Survey();
     survey.setValue("checkboxQuestion", "initialValue");
     var page = survey.addNewPage("my page");
@@ -152,7 +149,7 @@ QUnit.test(
     assert.deepEqual(question.value, ["setFromSurvey"], "set from survey");
   }
 );
-QUnit.test("Question Matrix: koValue in MatrixValue", function(assert) {
+QUnit.test("Question Matrix: koValue in MatrixValue", function (assert) {
   var matrix = new QuestionMatrix("q1");
   matrix.rows = ["row1", "row2"];
   matrix.columns = ["col1", "col2"];
@@ -167,7 +164,7 @@ QUnit.test("Question Matrix: koValue in MatrixValue", function(assert) {
     "the matrix value changed correctly"
   );
 });
-QUnit.test("Question Matrix: update koVisibleColumns on assign", function(
+QUnit.test("Question Matrix: update koVisibleColumns on assign", function (
   assert
 ) {
   var matrix = new QuestionMatrix("q1");
@@ -184,7 +181,7 @@ QUnit.test("Question Matrix: update koVisibleColumns on assign", function(
 });
 QUnit.test(
   "Question Matrix: change matrix value after visibleRows generation",
-  function(assert) {
+  function (assert) {
     var matrix = new QuestionMatrix("q1");
     matrix.rows = ["row1", "row2"];
     matrix.columns = ["col1", "col2"];
@@ -201,7 +198,7 @@ QUnit.test(
 
 QUnit.test(
   "Question Matrix: columnsVisibleIf, rowsVisibleIf, Bug #2012",
-  function(assert) {
+  function (assert) {
     var survey = new Survey();
     var page = survey.addNewPage("page1");
     var matrix = new QuestionMatrix("q1");
@@ -247,7 +244,7 @@ QUnit.test(
 
 QUnit.test(
   "Question MatrixDropdown: change matrix value after visibleRows generation",
-  function(assert) {
+  function (assert) {
     var matrix = new QuestionMatrixDropdown("q1");
     matrix.rows = ["row1", "row2", "row3"];
     matrix.columns.push(new MatrixDropdownColumn("column1"));
@@ -260,7 +257,7 @@ QUnit.test(
   }
 );
 
-QUnit.test("Matrixdynamic checkbox column does not work, Bug#1031", function(
+QUnit.test("Matrixdynamic checkbox column does not work, Bug#1031", function (
   assert
 ) {
   var survey = new Survey({
@@ -291,7 +288,7 @@ QUnit.test("Matrixdynamic checkbox column does not work, Bug#1031", function(
   );
 });
 
-QUnit.test("Question MultipleText: koValue in TextItem", function(assert) {
+QUnit.test("Question MultipleText: koValue in TextItem", function (assert) {
   var mQuestion = new QuestionMultipleText("q1");
   mQuestion.items.push(new MultipleTextItem("i1"));
   mQuestion.items.push(new MultipleTextItem("i2"));
@@ -315,7 +312,7 @@ QUnit.test("Question MultipleText: koValue in TextItem", function(assert) {
   mQuestion.value = null;
   assert.equal(mQuestion.items[0].value, null, "empty the value");
 });
-QUnit.test("Question MultipleText: koRows", function(assert) {
+QUnit.test("Question MultipleText: koRows", function (assert) {
   var mQuestion = new QuestionMultipleText("q1");
   mQuestion.items.push(new MultipleTextItem("i1"));
   mQuestion.items.push(new MultipleTextItem("i2"));
@@ -330,7 +327,7 @@ QUnit.test("Question MultipleText: koRows", function(assert) {
     "just one item in the first row"
   );
 });
-QUnit.test("koElements", function(assert) {
+QUnit.test("koElements", function (assert) {
   var survey = new Survey();
   var page = survey.addNewPage("page1");
   page.addNewQuestion("text", "q1");
@@ -353,7 +350,7 @@ QUnit.test("koElements", function(assert) {
     "One element in the second row, ko"
   );
 });
-QUnit.test("Set notification on setting survey data", function(assert) {
+QUnit.test("Set notification on setting survey data", function (assert) {
   var survey = new Survey();
   var page = survey.addNewPage("page1");
   var question = page.addNewQuestion("text", "q1");
@@ -362,7 +359,7 @@ QUnit.test("Set notification on setting survey data", function(assert) {
   assert.equal(survey.getValue("q1"), "value2", "survey data for q1");
   assert.equal(question.value, "value2", "knockout value is updated.");
 });
-QUnit.test("On make survey data empy for Multiple text question", function(
+QUnit.test("On make survey data empy for Multiple text question", function (
   assert
 ) {
   var survey = new Survey();
@@ -375,7 +372,7 @@ QUnit.test("On make survey data empy for Multiple text question", function(
   survey.data = null;
   assert.equal(question.items[0].value, null, "Make the data empty");
 });
-QUnit.test("isVisible property", function(assert) {
+QUnit.test("isVisible property", function (assert) {
   var survey = new Survey();
   var page = survey.addNewPage("page1");
   var question = page.addNewQuestion("text", "q1");
@@ -383,7 +380,7 @@ QUnit.test("isVisible property", function(assert) {
   question.visible = false;
   assert.equal(question.isVisible, false, "it is false now");
 });
-QUnit.test("comment property", function(assert) {
+QUnit.test("comment property", function (assert) {
   var survey = new Survey();
   survey.data = { q: "other", "q-Comment": "aaaa" };
   var page = survey.addNewPage("page1");
@@ -393,7 +390,7 @@ QUnit.test("comment property", function(assert) {
   question.hasOther = true;
   assert.equal(question.comment, "aaaa", "Set ko Comment");
 });
-QUnit.test("Load title correctly from JSON", function(assert) {
+QUnit.test("Load title correctly from JSON", function (assert) {
   var survey = new Survey({ questions: [{ type: "text", name: "question1" }] });
   assert.equal(
     survey.pages[0].questions[0]["locTitle"]["koRenderedHtml"](),
@@ -401,7 +398,7 @@ QUnit.test("Load title correctly from JSON", function(assert) {
     "title is getting from name"
   );
 });
-QUnit.test("errors should be empty after prevPage bug#151", function(assert) {
+QUnit.test("errors should be empty after prevPage bug#151", function (assert) {
   var survey = new Survey();
   survey.goNextPageAutomatic = true;
   var page = survey.addNewPage("page1");
@@ -425,7 +422,7 @@ QUnit.test("errors should be empty after prevPage bug#151", function(assert) {
   assert.equal(errorCount, 0, "The question has not errors");
 });
 
-QUnit.test("add customwidget item", function(assert) {
+QUnit.test("add customwidget item", function (assert) {
   CustomWidgetCollection.Instance.clear();
   CustomWidgetCollection.Instance.addCustomWidget({
     name: "first",
@@ -444,7 +441,7 @@ QUnit.test("add customwidget item", function(assert) {
   CustomWidgetCollection.Instance.clear();
 });
 
-QUnit.test("Localization, choices.locText.koRenderedHtml, #349", function(
+QUnit.test("Localization, choices.locText.koRenderedHtml, #349", function (
   assert
 ) {
   var survey = new Survey();
@@ -475,7 +472,7 @@ QUnit.test("Localization, choices.locText.koRenderedHtml, #349", function(
   survey.locale = "";
 });
 
-QUnit.test("Localization, otherItem", function(assert) {
+QUnit.test("Localization, otherItem", function (assert) {
   var survey = new Survey();
   var page = survey.addNewPage("page");
   var q1 = <QuestionCheckbox>page.addNewQuestion("checkbox", "q1");
@@ -496,7 +493,7 @@ QUnit.test("Localization, otherItem", function(assert) {
   survey.locale = "";
 });
 
-QUnit.test("otherItem, set text, editor: #90", function(assert) {
+QUnit.test("otherItem, set text, editor: #90", function (assert) {
   var survey = new Survey({
     questions: [
       {
@@ -513,7 +510,7 @@ QUnit.test("otherItem, set text, editor: #90", function(assert) {
   assert.equal(q1.visibleChoices[2].text, "my other", "use otherText");
 });
 
-QUnit.test("Update page.title correctly with numbers", function(assert) {
+QUnit.test("Update page.title correctly with numbers", function (assert) {
   var survey = new Survey();
   survey.addPage(createPageWithQuestion("page1"));
   survey.addPage(createPageWithQuestion("page2"));
@@ -536,7 +533,7 @@ QUnit.test("Update page.title correctly with numbers", function(assert) {
 
 QUnit.test(
   "Survey display mode should set isReadOnly to true for questions",
-  function(assert) {
+  function (assert) {
     var survey = new Survey();
     var page = new Page("page1");
     survey.addPage(page);
@@ -584,7 +581,7 @@ QUnit.test(
 
 QUnit.test(
   "Matrixdynamic adjust rowCount on setting the survey.data with another locale",
-  function(assert) {
+  function (assert) {
     var survey = new Survey();
     survey.addNewPage("p1");
     var question = new QuestionMatrixDynamic("q1");
@@ -598,7 +595,7 @@ QUnit.test(
   }
 );
 
-QUnit.test("Text preprocessing variable and value. Fix the bug#461", function(
+QUnit.test("Text preprocessing variable and value. Fix the bug#461", function (
   assert
 ) {
   var survey = new Survey();
@@ -628,7 +625,7 @@ QUnit.test("Text preprocessing variable and value. Fix the bug#461", function(
   );
 });
 
-QUnit.test("Load Panel from Json + visibleIf + startWithNewLine", function(
+QUnit.test("Load Panel from Json + visibleIf + startWithNewLine", function (
   assert
 ) {
   var json = {
@@ -662,7 +659,7 @@ QUnit.test("Load Panel from Json + visibleIf + startWithNewLine", function(
   );
 });
 
-QUnit.test("Load Panel from Json + isSinglePage", function(assert) {
+QUnit.test("Load Panel from Json + isSinglePage", function (assert) {
   var json = {
     isSinglePage: true,
     pages: [
@@ -688,7 +685,7 @@ QUnit.test("Load Panel from Json + isSinglePage", function(assert) {
   assert.equal(q.isVisible, true, "The question is visible");
 });
 
-QUnit.test("Load PanelDynamic from Json", function(assert) {
+QUnit.test("Load PanelDynamic from Json", function (assert) {
   var json = {
     questions: [
       {
@@ -745,7 +742,7 @@ QUnit.test("Load PanelDynamic from Json", function(assert) {
   assert.equal(question.panels.length, 2, "2 panels, panels has been updated");
 });
 
-QUnit.test("Load PanelDynamic from Json, nested panel", function(assert) {
+QUnit.test("Load PanelDynamic from Json, nested panel", function (assert) {
   var json = {
     questions: [
       {
@@ -826,7 +823,7 @@ QUnit.test("Load PanelDynamic from Json, nested panel", function(assert) {
   );
 });
 
-QUnit.test("PanelDynamic and koRenderedHtml on text processing", function(
+QUnit.test("PanelDynamic and koRenderedHtml on text processing", function (
   assert
 ) {
   var json = {
@@ -864,7 +861,7 @@ QUnit.test("PanelDynamic and koRenderedHtml on text processing", function(
   assert.equal(pLocTitle["koRenderedHtml"](), "val1", "np1 title is q1.value");
 });
 
-QUnit.test("loc strings changed only once on data assignment", function(
+QUnit.test("loc strings changed only once on data assignment", function (
   assert
 ) {
   var json = {
@@ -878,7 +875,7 @@ QUnit.test("loc strings changed only once on data assignment", function(
   let callCount = 0;
   let calls = "";
 
-  survey.onTextMarkdown.add(function(survey, options) {
+  survey.onTextMarkdown.add(function (survey, options) {
     callCount++;
     calls += "->" + options.element.name + "." + options.name;
   });
@@ -908,7 +905,7 @@ QUnit.test("loc strings changed only once on data assignment", function(
   );
 });
 
-QUnit.test("koSurvey matrix.rowsVisibleIf", function(assert) {
+QUnit.test("koSurvey matrix.rowsVisibleIf", function (assert) {
   var survey = new Survey();
   var page = survey.addNewPage("p1");
   var qCars = new QuestionCheckbox("cars");
@@ -944,7 +941,7 @@ export class DesignerSurveyTester extends Survey {
 
 QUnit.test(
   "Questions are not rendered in Panel inside Dynamic Panel in Editor, editor bug #216",
-  function(assert) {
+  function (assert) {
     var survey = new DesignerSurveyTester({
       pages: [
         {
@@ -996,7 +993,7 @@ QUnit.test(
 
 QUnit.test(
   "Add invisible question in Panel from JSON in Editor, editor bug #218",
-  function(assert) {
+  function (assert) {
     var survey = new DesignerSurveyTester();
     var page = survey.addNewPage("page1");
     var json = {
@@ -1029,7 +1026,7 @@ QUnit.test(
 
 QUnit.test(
   "Load rating from JSON, bug# https://github.com/surveyjs/editor/issues/171",
-  function(assert) {
+  function (assert) {
     var json = {
       questions: [
         { type: "rating", name: "q", rateMax: 8, rateMin: 2, rateStep: 2 },
@@ -1045,7 +1042,7 @@ QUnit.test(
     );
   }
 );
-QUnit.test("Default value doesn't set in PanelDynamic , bug#910", function(
+QUnit.test("Default value doesn't set in PanelDynamic , bug#910", function (
   assert
 ) {
   var json = {
@@ -1080,7 +1077,7 @@ QUnit.test("Default value doesn't set in PanelDynamic , bug#910", function(
 
 QUnit.test(
   "Survey.isSinglePage = true, the last page doesn't added, bug#1009",
-  function(assert) {
+  function (assert) {
     var json = {
       pages: [
         {
@@ -1132,7 +1129,7 @@ QUnit.test(
   }
 );
 
-QUnit.test("Survey Localization - radiogroup.otheItem, Bug#1045", function(
+QUnit.test("Survey Localization - radiogroup.otheItem, Bug#1045", function (
   assert
 ) {
   var json = {
@@ -1181,7 +1178,7 @@ QUnit.test("Survey Localization - radiogroup.otheItem, Bug#1045", function(
 
 QUnit.test(
   "PanelDynamic and MatrixDynamic, survey in readonly mode, Bug#1051",
-  function(assert) {
+  function (assert) {
     var json = {
       questions: [
         {
@@ -1251,7 +1248,7 @@ function createPageWithQuestion(name: string): Page {
   return page;
 }
 
-QUnit.test("koquestion inside panel vidibleif", function(assert) {
+QUnit.test("koquestion inside panel vidibleif", function (assert) {
   var json = {
     pages: [
       {
@@ -1310,7 +1307,7 @@ QUnit.test("koquestion inside panel vidibleif", function(assert) {
 
 QUnit.test(
   "multipletext item is not readonly when survey is readonly, bug #1177",
-  function(assert) {
+  function (assert) {
     var json = {
       mode: "display",
       elements: [
@@ -1333,7 +1330,7 @@ QUnit.test(
   }
 );
 
-QUnit.test("Dynamic Panel bug with localization, bug #1184", function(assert) {
+QUnit.test("Dynamic Panel bug with localization, bug #1184", function (assert) {
   var json = {
     locale: "de",
     elements: [
@@ -1429,7 +1426,7 @@ QUnit.test("Dynamic Panel bug with localization, bug #1184", function(assert) {
 
 QUnit.test(
   "exception during changing multiSelect for imagepicker, bug https://github.com/surveyjs/editor/issues/374",
-  function(assert) {
+  function (assert) {
     var survey = new Survey();
     var page = survey.addNewPage("p");
     var q = new QuestionImagePicker("question1");
@@ -1443,7 +1440,7 @@ QUnit.test(
   }
 );
 
-QUnit.test("Could not assign value into mutlipletext question, #1229", function(
+QUnit.test("Could not assign value into mutlipletext question, #1229", function (
   assert
 ) {
   var survey = new Survey();
@@ -1465,7 +1462,7 @@ QUnit.test("Could not assign value into mutlipletext question, #1229", function(
   );
 });
 
-QUnit.test("Checkbox Select All Test", function(assert) {
+QUnit.test("Checkbox Select All Test", function (assert) {
   var survey = new Survey();
   var page = survey.addNewPage("page1");
   var question = new QuestionCheckbox("q1");
@@ -1486,7 +1483,7 @@ QUnit.test("Checkbox Select All Test", function(assert) {
 
 QUnit.test(
   "Changing isRequired doesn't update title for questions in dynamic panel, Bug in Editor #385",
-  function(assert) {
+  function (assert) {
     var survey = new Survey();
     var page = survey.addNewPage("page1");
     var question = new QuestionPanelDynamic("q1");
@@ -1517,7 +1514,7 @@ QUnit.test(
 
 QUnit.test(
   "panel.visibleIf doesn't work if it is a single panel on the page, #1329",
-  function(assert) {
+  function (assert) {
     var json = {
       pages: [
         {
@@ -1611,7 +1608,7 @@ QUnit.test("Expression with two columns doesn't work, bug#1199", function(
 });
 */
 
-QUnit.test("Loc string text preprocessing update on change value", function(
+QUnit.test("Loc string text preprocessing update on change value", function (
   assert
 ) {
   var json = {
@@ -1655,7 +1652,7 @@ QUnit.test("Loc string text preprocessing update on change value", function(
     "The first value"
   );
 });
-QUnit.test("question.paddingLeft and question.paddingRight", function(assert) {
+QUnit.test("question.paddingLeft and question.paddingRight", function (assert) {
   var survey = new Survey({
     elements: [{ type: "panel", name: "p1" }],
   });
@@ -1669,7 +1666,7 @@ QUnit.test("question.paddingLeft and question.paddingRight", function(assert) {
   assert.equal(panel.paddingRight, "40px", "right is not empty");
   assert.equal(panel.paddingLeft, "60px", "left is not empty");
 });
-QUnit.test("panel and question paddings form json", function(assert) {
+QUnit.test("panel and question paddings form json", function (assert) {
   var survey = new Survey({
     elements: [
       {
@@ -1687,7 +1684,7 @@ QUnit.test("panel and question paddings form json", function(assert) {
   assert.equal(panel.innerPaddingLeft, "40px", "panel inner left is 40px");
   assert.equal(question.paddingLeft, "20px", "question left is 20px");
 });
-QUnit.test("Questions are randomized", function(assert) {
+QUnit.test("Questions are randomized", function (assert) {
   var survey = new Survey({
     elements: [
       {
@@ -1716,7 +1713,7 @@ function doTimer(count: number) {
 }
 QUnit.test(
   "Complete pages by timer and questionsOrder = 'random', Bug#1406, Bug#1482",
-  function(assert) {
+  function (assert) {
     var survey = new Survey({
       pages: [
         {
@@ -1764,7 +1761,7 @@ QUnit.test(
   }
 );
 
-QUnit.test("https://github.com/surveyjs/surveyjs/issues/1501", function(
+QUnit.test("https://github.com/surveyjs/surveyjs/issues/1501", function (
   assert
 ) {
   var survey = new Survey();
@@ -1787,7 +1784,7 @@ QUnit.test("https://github.com/surveyjs/surveyjs/issues/1501", function(
 
 QUnit.test(
   "Some custom CSS not working - https://github.com/surveyjs/survey-library/issues/2122",
-  function(assert) {
+  function (assert) {
     var survey = new Survey({
       questions: [
         {
@@ -1833,7 +1830,7 @@ QUnit.test(
 
 QUnit.test(
   "Other item selected and not checked - https://github.com/surveyjs/survey-library/issues/2200",
-  function(assert) {
+  function (assert) {
     var q1 = new QuestionRadiogroup("q1");
     q1.fromJSON({
       type: "radiogroup",
@@ -1864,7 +1861,7 @@ QUnit.test(
   }
 );
 
-QUnit.test("survey.firstPageIsStarted=true + multiple-language", function(
+QUnit.test("survey.firstPageIsStarted=true + multiple-language", function (
   assert
 ) {
   var json = {
@@ -1905,7 +1902,7 @@ QUnit.test("survey.firstPageIsStarted=true + multiple-language", function(
   survey.locale = prevLocale;
 });
 
-QUnit.test("survey.firstPageIsStarted=true + multiple-language", function(
+QUnit.test("survey.firstPageIsStarted=true + multiple-language", function (
   assert
 ) {
   var survey = new Survey({
@@ -1931,7 +1928,7 @@ QUnit.test("survey.firstPageIsStarted=true + multiple-language", function(
   assert.equal(survey.startedPage.name, "startedPage", "Loaded fine");
 });
 
-QUnit.test("ProgressButtonsViewModel component scroll button", function(
+QUnit.test("ProgressButtonsViewModel component scroll button", function (
   assert
 ) {
   let json: any = {
@@ -1969,7 +1966,7 @@ QUnit.test("ProgressButtonsViewModel component scroll button", function(
   let progress: ProgressButtonsViewModel = new ProgressButtonsViewModel(
     survey,
     {
-      querySelector: function() {
+      querySelector: function () {
         return undefined;
       },
     }
@@ -1978,15 +1975,15 @@ QUnit.test("ProgressButtonsViewModel component scroll button", function(
   assert.equal(
     progress.getScrollButtonCss(true)(),
     survey.css.progressButtonsImageButtonLeft +
-      " " +
-      survey.css.progressButtonsImageButtonHidden,
+    " " +
+    survey.css.progressButtonsImageButtonHidden,
     "1) Scroll button left style is hidden"
   );
   assert.equal(
     progress.getScrollButtonCss(false)(),
     survey.css.progressButtonsImageButtonRight +
-      " " +
-      survey.css.progressButtonsImageButtonHidden,
+    " " +
+    survey.css.progressButtonsImageButtonHidden,
     "1) Scroll button right style is hidden"
   );
 
@@ -2003,7 +2000,7 @@ QUnit.test("ProgressButtonsViewModel component scroll button", function(
   );
 });
 
-QUnit.test("QuestionCheckbox otherItem KO text", function(assert) {
+QUnit.test("QuestionCheckbox otherItem KO text", function (assert) {
   var survey = new Survey();
   survey.locale = "";
   survey.addNewPage();
@@ -2022,7 +2019,7 @@ QUnit.test("QuestionCheckbox otherItem KO text", function(assert) {
   );
 });
 
-QUnit.test("QuestionCheckbox otherItem KO text design time", function(assert) {
+QUnit.test("QuestionCheckbox otherItem KO text design time", function (assert) {
   var survey = new Survey();
   survey.locale = "";
   survey.setDesignMode(true);
@@ -2042,7 +2039,7 @@ QUnit.test("QuestionCheckbox otherItem KO text design time", function(assert) {
   );
 });
 
-QUnit.test("Question KO title changed on name", function(assert) {
+QUnit.test("Question KO title changed on name", function (assert) {
   var survey = new Survey();
   survey.locale = "";
   survey.addNewPage();
@@ -2064,7 +2061,7 @@ QUnit.test("Question KO title changed on name", function(assert) {
 });
 QUnit.test(
   "Text processing in rows and columns, rendered table, Bug#2829",
-  function(assert) {
+  function (assert) {
     var survey = new Survey({
       elements: [
         {
@@ -2117,7 +2114,7 @@ QUnit.test(
 );
 QUnit.test(
   "check imageLink in imagePicker items on changing location, Bug#3287",
-  function(assert) {
+  function (assert) {
     const survey = new Survey({
       elements: [
         {
@@ -2144,7 +2141,7 @@ QUnit.test(
   }
 );
 
-QUnit.test("Change title koRenderedHtml on changing name", function(assert) {
+QUnit.test("Change title koRenderedHtml on changing name", function (assert) {
   var survey = new Survey({
     elements: [
       {
@@ -2166,11 +2163,11 @@ QUnit.test("Change title koRenderedHtml on changing name", function(assert) {
     "The  value is changed"
   );
 });
-QUnit.test("Survey Markdown, Bug:#2831", function(assert) {
+QUnit.test("Survey Markdown, Bug:#2831", function (assert) {
   var survey = new Survey({
     elements: [{ type: "text", name: "q1", title: "Hellomarkdown" }],
   });
-  survey.onTextMarkdown.add(function(survey, options) {
+  survey.onTextMarkdown.add(function (survey, options) {
     if (options.text.indexOf("markdown") > -1)
       options.html = options.text.replace("markdown", "!");
   });
@@ -2181,7 +2178,7 @@ QUnit.test("Survey Markdown, Bug:#2831", function(assert) {
     "Process markdown correctly"
   );
 });
-QUnit.test("Call loc str changed onGetQuestionTitle, Bug#3080", function(
+QUnit.test("Call loc str changed onGetQuestionTitle, Bug#3080", function (
   assert
 ) {
   var survey = new Survey({
@@ -2203,7 +2200,7 @@ QUnit.test("Call loc str changed onGetQuestionTitle, Bug#3080", function(
     "Process onGetQuestionTitle correctly"
   );
 });
-QUnit.test("loc strings changed on data assignment", function(assert) {
+QUnit.test("loc strings changed on data assignment", function (assert) {
   var json = {
     questions: [
       { type: "text", name: "q1", title: "Question 1" },
@@ -2230,7 +2227,7 @@ QUnit.test("loc strings changed on data assignment", function(assert) {
 });
 QUnit.test(
   "Create correct Panel object when creating in code, Bug#2866",
-  function(assert) {
+  function (assert) {
     var survey = new Survey();
     var page = survey.addNewPage("page1");
     var panel = <Panel>page.addNewPanel("panel1");
@@ -2270,7 +2267,7 @@ QUnit.test(
 );
 QUnit.test(
   "start page and boolean question, check that locLabelTrue/False set correctly, Bug#2947",
-  function(assert) {
+  function (assert) {
     var survey = new Survey({
       firstPageIsStarted: true,
       pages: [
@@ -2310,7 +2307,7 @@ QUnit.test(
 );
 QUnit.test(
   "MatrixDynamic, test renderedTable column locString on adding new column",
-  function(assert) {
+  function (assert) {
     var survey = new Survey({
       elements: [
         {
@@ -2347,7 +2344,7 @@ QUnit.test(
 );
 QUnit.test(
   "MatrixDynamic, test renderedTable column locString for checkbox question",
-  function(assert) {
+  function (assert) {
     var survey = new Survey({
       elements: [
         {
@@ -2390,7 +2387,7 @@ QUnit.test(
 );
 QUnit.test(
   "MatrixDynamic, test renderedTable column locString after changing default cell Type",
-  function(assert) {
+  function (assert) {
     var survey = new Survey({
       elements: [
         {
@@ -2419,7 +2416,7 @@ QUnit.test(
     );
   }
 );
-QUnit.test("Error on running triggers", function(assert) {
+QUnit.test("Error on running triggers", function (assert) {
   var survey = new Survey({
     elements: [
       {
@@ -2467,7 +2464,7 @@ QUnit.test("Error on running triggers", function(assert) {
   survey.setValue("matrix", [{ col1: "v" }]);
   assert.equal(survey.getValue("radio"), "no", "Set value on trigger");
 });
-QUnit.test("pageNext/pagePrev... Text and koRenderedHtml", function(assert) {
+QUnit.test("pageNext/pagePrev... Text and koRenderedHtml", function (assert) {
   var survey = new Survey({
     elements: [
       {
@@ -2480,7 +2477,7 @@ QUnit.test("pageNext/pagePrev... Text and koRenderedHtml", function(assert) {
   survey.completeText = "Finish";
   assert.equal((<any>survey.locCompleteText).koRenderedHtml(), "Finish");
 });
-QUnit.test("Initial Text Processing in panel title and ko", function(assert) {
+QUnit.test("Initial Text Processing in panel title and ko", function (assert) {
   var survey = new Survey({
     elements: [
       {
