+++ tests/lowercasetests.ts
@@ -2,33 +2,31 @@ import { SurveyModel } from "../src/survey";
 import { PageModel } from "../src/page";
 import { QuestionTextModel } from "../src/question_text";
 import { MultipleTextItemModel } from "../src/question_multipletext";
-import {
-  QuestionMatrixDropdownModelBase,
-  MatrixDropdownColumn
-} from "../src/question_matrixdropdownbase";
+import { QuestionMatrixDropdownModelBase } from "../src/question_matrixdropdownbase";
 import { QuestionSelectBase } from "../src/question_baseselect";
+import { MatrixDropdownColumn } from "../src/question_matrixdropdowncolumn";
 
 export default QUnit.module("SurveyLowercase");
 
-QUnit.test("inputType value is always lower-case", function(assert) {
+QUnit.test("inputType value is always lower-case", function (assert) {
   var question = new QuestionTextModel("text");
   question.inputType = "TEXT";
   assert.strictEqual(question.inputType, "text");
 });
 
-QUnit.test("inputType value is always lower-case", function(assert) {
+QUnit.test("inputType value is always lower-case", function (assert) {
   var question = new QuestionTextModel("text");
   question.inputType = "DATETIME_LOCAL";
   assert.strictEqual(question.inputType, "datetime-local");
 });
 
-QUnit.test("choicesOrder value is always lower-case", function(assert) {
+QUnit.test("choicesOrder value is always lower-case", function (assert) {
   var question = new QuestionSelectBase("base");
   question.choicesOrder = "RANDOM";
   assert.strictEqual(question.choicesOrder, "random");
 });
 
-QUnit.test("navigationButtonsVisibility value is always lower-case", function(
+QUnit.test("navigationButtonsVisibility value is always lower-case", function (
   assert
 ) {
   var question = new PageModel("base");
@@ -38,7 +36,7 @@ QUnit.test("navigationButtonsVisibility value is always lower-case", function(
 
 QUnit.test(
   "MatrixDropdownColumn inputType value is always lower-case",
-  function(assert) {
+  function (assert) {
     var question = new MatrixDropdownColumn("text");
     question.cellType = "text";
     question["inputType"] = "TEXT";
@@ -46,7 +44,7 @@ QUnit.test(
   }
 );
 
-QUnit.test("MatrixDropdownColumn cellType value is always lower-case", function(
+QUnit.test("MatrixDropdownColumn cellType value is always lower-case", function (
   assert
 ) {
   var question = new MatrixDropdownColumn("base");
@@ -56,7 +54,7 @@ QUnit.test("MatrixDropdownColumn cellType value is always lower-case", function(
 
 QUnit.test(
   "MatrixDropdownColumn choicesOrder value is always lower-case",
-  function(assert) {
+  function (assert) {
     var question = new MatrixDropdownColumn("base");
     question["choicesOrder"] = "RANDOM";
     assert.strictEqual(question["choicesOrder"], "random");
@@ -65,7 +63,7 @@ QUnit.test(
 
 QUnit.test(
   "QuestionMatrixDropdownModelBase cellType value is always lower-case",
-  function(assert) {
+  function (assert) {
     var question = new QuestionMatrixDropdownModelBase("base");
     question.cellType = "RADIOGROUP";
     assert.strictEqual(question.cellType, "radiogroup");
@@ -74,7 +72,7 @@ QUnit.test(
 
 QUnit.test(
   "MultipleTextItemModel inputType value is always lower-case",
-  function(assert) {
+  function (assert) {
     var question = new MultipleTextItemModel("text");
     question.inputType = "COLOR";
     assert.strictEqual(question.inputType, "color");
@@ -83,7 +81,7 @@ QUnit.test(
 
 QUnit.test(
   "SurveyModel showQuestionNumbers value is always lower-case",
-  function(assert) {
+  function (assert) {
     var survey = new SurveyModel();
     survey.showQuestionNumbers = "OFF";
     assert.strictEqual(survey.showQuestionNumbers, "off");
@@ -92,7 +90,7 @@ QUnit.test(
 
 QUnit.test(
   "SurveyModel showQuestionNumbers value handles onPage special case",
-  function(assert) {
+  function (assert) {
     var survey = new SurveyModel();
     survey.showQuestionNumbers = "ONPAGE";
     assert.strictEqual(survey.showQuestionNumbers, "onPage");
@@ -101,14 +99,14 @@ QUnit.test(
 
 QUnit.test(
   "SurveyModel questionTitleLocation value is always lower-case",
-  function(assert) {
+  function (assert) {
     var survey = new SurveyModel();
     survey.questionTitleLocation = "BOTTOM";
     assert.strictEqual(survey.questionTitleLocation, "bottom");
   }
 );
 
-QUnit.test("SurveyModel showProgressBar value is always lower-case", function(
+QUnit.test("SurveyModel showProgressBar value is always lower-case", function (
   assert
 ) {
   var survey = new SurveyModel();
@@ -117,7 +115,7 @@ QUnit.test("SurveyModel showProgressBar value is always lower-case", function(
   assert.equal(survey.isShowProgressBarOnTop, true);
 });
 
-QUnit.test("SurveyModel mode value is always lower-case", function(assert) {
+QUnit.test("SurveyModel mode value is always lower-case", function (assert) {
   var survey = new SurveyModel();
   survey.mode = "DISPLAY";
   assert.strictEqual(survey.mode, "display");
