+++ tests/question_matrixdropdownbasetests.ts
@@ -1,17 +1,20 @@
-import {
-  QuestionMatrixDropdownModelBase,
-  MatrixDropdownColumn
-} from "../src/question_matrixdropdownbase";
+import { QuestionMatrixDropdownModelBase } from "../src/question_matrixdropdownbase";
 
 export default QUnit.module("Survey_QuestionMatrixDropdownBase");
 
-QUnit.test(" 213 123sad d213 sadsd", function(assert) {
-  var matrix = new QuestionMatrixDropdownModelBase("q1");
+QUnit.test(" 213 123sad d213 sadsd", function (assert) {
+  const matrix = new QuestionMatrixDropdownModelBase("q1");
 
-  assert.equal(
-    matrix["generatedVisibleRows"],
-    null,
-    "generatedVisibleRows is null"
-  );
+  assert.equal(matrix["generatedVisibleRows"], null, "generatedVisibleRows is null");
   assert.deepEqual(matrix.getAllErrors(), [], "getAllErrors method doesn't fall");
+});
+QUnit.test("allowAdaptiveActions", function (assert) {
+  const matrix = new QuestionMatrixDropdownModelBase("q1");
+  assert.equal(matrix.allowAdaptiveActions, true, "matrix.allowAdaptiveActions");
+  assert.equal(matrix.getPanel()["allowAdaptiveActions"], true, "matrix.panel.allowAdaptiveActions");
+
+  matrix.allowAdaptiveActions = false;
+  assert.equal(matrix.allowAdaptiveActions, false, "matrix.allowAdaptiveActions");
+  assert.equal(matrix.getPanel()["allowAdaptiveActions"], false, "matrix.panel.allowAdaptiveActions");
+
 });
