+++ tests/surveyserializationtests.ts
@@ -4,23 +4,18 @@ import { PageModel } from "../src/page";
 import { QuestionTextModel } from "../src/question_text";
 import { QuestionCheckboxModel } from "../src/question_checkbox";
 import { Question } from "../src/question";
-import {
-  QuestionMultipleTextModel,
-  MultipleTextItemModel,
-} from "../src/question_multipletext";
+import { QuestionMultipleTextModel, MultipleTextItemModel } from "../src/question_multipletext";
 import { QuestionDropdownModel } from "../src/question_dropdown";
-import {
-  QuestionMatrixDropdownModelBase,
-  MatrixDropdownColumn,
-} from "../src/question_matrixdropdownbase";
+import { QuestionMatrixDropdownModelBase } from "../src/question_matrixdropdownbase";
 import { ItemValue } from "../src/itemvalue";
 import { NumericValidator } from "../src/validator";
 import { QuestionRadiogroupModel } from "../src/question_radiogroup";
 import { Helpers } from "../src/helpers";
+import { MatrixDropdownColumn } from "../src/question_matrixdropdowncolumn";
 
 export default QUnit.module("SurveySerialization");
 
-QUnit.test("Serialize two pages", function(assert) {
+QUnit.test("Serialize two pages", function (assert) {
   var survey = new SurveyModel();
   survey.addNewPage("Page 1");
   survey.addNewPage("Page 2");
@@ -31,7 +26,7 @@ QUnit.test("Serialize two pages", function(assert) {
     "serialize two pages"
   );
 });
-QUnit.test("Deserialize two pages", function(assert) {
+QUnit.test("Deserialize two pages", function (assert) {
   var survey = new SurveyModel();
   new JsonObject().toObject(
     { pages: [{ name: "Page1" }, { name: "Page2" }] },
@@ -42,7 +37,7 @@ QUnit.test("Deserialize two pages", function(assert) {
   assert.equal(survey.pages[0].data, survey, "data interface is set");
   assert.equal(survey.pages[1].getType(), "page", "it is a live object");
 });
-QUnit.test("Serialize two questions", function(assert) {
+QUnit.test("Serialize two questions", function (assert) {
   var page = new PageModel("Page1");
   var textQuestion = new QuestionTextModel("textQuestion");
   textQuestion.isRequired = true;
@@ -59,7 +54,7 @@ QUnit.test("Serialize two questions", function(assert) {
     "serialize two questions"
   );
 });
-QUnit.test("Deserialize two questions", function(assert) {
+QUnit.test("Deserialize two questions", function (assert) {
   var survey = new SurveyModel();
   var page = new PageModel("Page1");
   survey.addPage(page);
@@ -113,7 +108,7 @@ QUnit.test("Deserialize two questions", function(assert) {
     "data interface is working"
   );
 });
-QUnit.test("Full survey deserialize with one question", function(assert) {
+QUnit.test("Full survey deserialize with one question", function (assert) {
   var survey = new SurveyModel();
   new JsonObject().toObject(
     {
@@ -143,7 +138,7 @@ QUnit.test("Full survey deserialize with one question", function(assert) {
 });
 QUnit.test(
   "Full survey deserialize with one question bypass pages object",
-  function(assert) {
+  function (assert) {
     var survey = new SurveyModel();
     new JsonObject().toObject(
       {
@@ -169,7 +164,7 @@ QUnit.test(
 );
 QUnit.test(
   "Full survey deserialize with one element bypass pages object",
-  function(assert) {
+  function (assert) {
     var survey = new SurveyModel();
     new JsonObject().toObject(
       {
@@ -193,7 +188,7 @@ QUnit.test(
     );
   }
 );
-QUnit.test("Serialize survey data", function(assert) {
+QUnit.test("Serialize survey data", function (assert) {
   var survey = new SurveyModel();
   survey.setValue("question1", "value1");
   survey.setValue("question2", true);
@@ -206,7 +201,7 @@ QUnit.test("Serialize survey data", function(assert) {
   };
   assert.deepEqual(data, expectedData, "check if get data works correctly");
 });
-QUnit.test("Deserialize survey data", function(assert) {
+QUnit.test("Deserialize survey data", function (assert) {
   var survey = new SurveyModel();
   var data = {
     question1: "value1",
@@ -230,7 +225,7 @@ QUnit.test("Deserialize survey data", function(assert) {
     "survey data for question 3"
   );
 });
-QUnit.test("Serialize mutltiple text question", function(assert) {
+QUnit.test("Serialize mutltiple text question", function (assert) {
   var mtQuestion = new QuestionMultipleTextModel("q1");
   mtQuestion.items.push(new MultipleTextItemModel("item1"));
   mtQuestion.items.push(new MultipleTextItemModel("item2", "text2"));
@@ -243,7 +238,7 @@ QUnit.test("Serialize mutltiple text question", function(assert) {
 });
 QUnit.test(
   "Deserialize/serialize mutltiple text question default value",
-  function(assert) {
+  function (assert) {
     var json = {
       name: "q",
       defaultValue: { item1: "11", item2: "22" },
@@ -266,7 +261,7 @@ QUnit.test(
     assert.deepEqual(q.toJSON(), json, "Default value serialized correctly");
   }
 );
-QUnit.test("Serialize restful choices", function(assert) {
+QUnit.test("Serialize restful choices", function (assert) {
   var question = new QuestionDropdownModel("q1");
   question.choicesByUrl.path = "name";
   var jsObj = new JsonObject().toJsonObject(question);
@@ -276,7 +271,7 @@ QUnit.test("Serialize restful choices", function(assert) {
     "serialize choicesByUrl"
   );
 });
-QUnit.test("Deserialize question with missing name", function(assert) {
+QUnit.test("Deserialize question with missing name", function (assert) {
   var survey = new SurveyModel();
   var jsonObj = new JsonObject();
   jsonObj.toObject(
@@ -297,7 +292,7 @@ QUnit.test("Deserialize question with missing name", function(assert) {
     "The required property error"
   );
 });
-QUnit.test("Deserialize choicesByUrl", function(assert) {
+QUnit.test("Deserialize choicesByUrl", function (assert) {
   var question = new QuestionDropdownModel("q1");
   assert.equal(
     question.choicesByUrl.isEmpty,
@@ -328,7 +323,7 @@ QUnit.test("Deserialize choicesByUrl", function(assert) {
     "data is copied correctly"
   );
 });
-QUnit.test("MatrixDropdown serialize and deserialize", function(assert) {
+QUnit.test("MatrixDropdown serialize and deserialize", function (assert) {
   var matrix = new QuestionMatrixDropdownModelBase("q1");
   matrix.columns.push(new MatrixDropdownColumn("col1"));
   matrix.columns.push(new MatrixDropdownColumn("col2"));
@@ -345,7 +340,7 @@ QUnit.test("MatrixDropdown serialize and deserialize", function(assert) {
   );
 });
 
-QUnit.test("Survey serialize dropdown.choices localization", function(assert) {
+QUnit.test("Survey serialize dropdown.choices localization", function (assert) {
   var survey = new SurveyModel();
   var page = survey.addNewPage("page1");
   var q1 = <QuestionDropdownModel>page.addNewQuestion("dropdown", "question1");
@@ -374,7 +369,7 @@ QUnit.test("Survey serialize dropdown.choices localization", function(assert) {
   assert.deepEqual(json, checkedJson, "Jsons should be the same");
 });
 
-QUnit.test("Survey deserialize checkbox.choices localization", function(
+QUnit.test("Survey deserialize checkbox.choices localization", function (
   assert
 ) {
   //{ pages: [{elements: [], name: "page1"}]}
@@ -396,7 +391,7 @@ QUnit.test("Survey deserialize checkbox.choices localization", function(
   );
 });
 
-QUnit.test("Survey deserialize/serialize localization survey", function(
+QUnit.test("Survey deserialize/serialize localization survey", function (
   assert
 ) {
   var survey = new SurveyModel();
@@ -426,7 +421,7 @@ QUnit.test("Survey deserialize/serialize localization survey", function(
 
 QUnit.test(
   "Survey deserialize dynamic matrix with different locale, Issue #507",
-  function(assert) {
+  function (assert) {
     var survey = new SurveyModel({
       pages: [
         {
@@ -453,7 +448,7 @@ QUnit.test(
 
 QUnit.test(
   "Survey checkbox.choices serialize/deserialize custom properties",
-  function(assert) {
+  function (assert) {
     Serializer.addProperty("itemvalue", "imageLink");
     var question = new QuestionCheckboxModel("q1");
     var jsonObj = new JsonObject();
@@ -481,7 +476,7 @@ QUnit.test(
     Serializer.removeProperty("itemvalue", "imageLink");
   }
 );
-QUnit.test("Serialize numeric validation, minValue=0, Editor#239", function(
+QUnit.test("Serialize numeric validation, minValue=0, Editor#239", function (
   assert
 ) {
   var survey = new SurveyModel();
@@ -504,7 +499,7 @@ QUnit.test("Serialize numeric validation, minValue=0, Editor#239", function(
 
 QUnit.test(
   "Expressions + markup https://surveyjs.answerdesk.io/ticket/details/T909",
-  function(assert) {
+  function (assert) {
     var json = {
       pages: [
         {
@@ -533,7 +528,7 @@ QUnit.test(
     var survey = new SurveyModel(json);
     var q1 = survey.getQuestionByName("Q1");
     var q2: QuestionRadiogroupModel = <any>survey.getQuestionByName("Radio1");
-    survey.onTextMarkdown.add(function(survey, options) {
+    survey.onTextMarkdown.add(function (survey, options) {
       options.html = options.text;
     });
 
@@ -544,7 +539,7 @@ QUnit.test(
   }
 );
 
-QUnit.test("Do not deseriabled default comment text - 'Other text:'", function(
+QUnit.test("Do not deseriabled default comment text - 'Other text:'", function (
   assert
 ) {
   var survey = new SurveyModel();
@@ -576,7 +571,7 @@ QUnit.test("Do not deseriabled default comment text - 'Other text:'", function(
   );
 });
 
-QUnit.test("Test fromJSON/clone/ensureUniqueNames functionalities", function(
+QUnit.test("Test fromJSON/clone/ensureUniqueNames functionalities", function (
   assert
 ) {
   var survey = new SurveyModel();
@@ -638,7 +633,7 @@ QUnit.test("Test fromJSON/clone/ensureUniqueNames functionalities", function(
 
 QUnit.test(
   "html and expression questions should not have errors, Bug#2359",
-  function(assert) {
+  function (assert) {
     var survey = new SurveyModel({
       questions: [
         { name: "q1", type: "html", html: "text", isRequired: true },
